### 2019-11-08 #1878 @whedon
[PRE REVIEW]: SHARPy: A dynamic aeroelastic simulation toolbox for very flexible aircraft and wind turbines
**Submitting author:** @fonsocarre (<a href="http://orcid.org/0000-0002-8133-9481">Alfonso del Carre</a>)
**Repository:** <a href="https://github.com/ImperialCollegeLondon/sharpy" target ="_blank">https://github.com/ImperialCollegeLondon/sharpy</a>
**Version:** v1.0.0
