### 2019-11-09 #1878 @labarba
👋  @mtezzele — I see that you have published [three papers](https://joss.theoj.org/papers/by/Marco%20Tezzele) with JOSS. Good job!
And you also already [reviewed](https://joss.theoj.org/papers/reviewed_by/@mtezzele) one paper. Thank you 🙏

Could you contribute one more review for us? The submission is:
*SHARPy: A dynamic aeroelastic simulation toolbox for very flexible aircraft and wind turbines*
https://github.com/ImperialCollegeLondon/sharpy
