### 2019-11-09 #1878 @labarba
👋  @petebachant — Do you think you could contribute a review for JOSS on this submission? You haven't reviewed for us before, but I think you will like it. The whole process is done in the open, on a GitHub issue. Have a look at our [Reviewer Guide](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
The submission is:
_SHARPy: A dynamic aeroelastic simulation toolbox for very flexible aircraft and wind turbines_
https://github.com/ImperialCollegeLondon/sharpy
