### 2019-11-09 #1878 @labarba
👋 @rafmudaf — I found that you are a contributor of [OpenFAST](https://github.com/OpenFAST/openfast) at NREL. Would you be interested in contributing a review for [The Journal of Open Source Software](https://joss.theoj.org)? Reviews are done in the open, in a GitHub repo, and they focus on the quality of software for research uses. Have a look at our [Reviewer Guide](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The submission I'm hoping you'd like to review is:
_SHARPy: A dynamic aeroelastic simulation toolbox for very flexible aircraft and wind turbines_
https://github.com/ImperialCollegeLondon/sharpy
https://github.com/openjournals/joss-reviews/issues/1878
