### 2019-11-11 #1885 @labarba
👋  @petebachant, @rafmudaf — Thank you for agreeing to review for JOSS! This is where the action happens: work your way through the review checklist, feel free to ask questions or post comments here, and also open issues in the submission repository as needed (mentioning this issue, for a cross-link). Godspeed!
