### 2019-11-12 55098d7
Updated geradin tutorial in the install instructions + cleaned docignore
