+++ docs/source/content/installation.md
@@ -1,5 +1,5 @@
 # SHARPy Installation Guide
-__Last revision 13 December 2018__
+__Last revision 12 November 2019__
 
 The following step by step tutorial will guide you through the installation process of SHARPy.
 
@@ -233,14 +233,25 @@ You are now ready to run SHARPy cases from the terminal.
 
 ### The SHARPy Case Structure
 
+__Setting up a SHARPy case__
+
 SHARPy cases are usually structured in the following way:
 
-1. The `generate_case.py` file: contains the setup of the problem, like the geometry, flight conditions etc.
-This script creates two output files that will then be used by SHARPy, `.fem.h5` and the `.solver.txt` file.
+1. A `generate_case.py` file: contains the setup of the problem, like geometry, flight conditions etc.
+This script creates the output files that will then be used by SHARPy, namely:
+    * The structural `.fem.h5` file.
+    * The aerodynamic `.aero.h5` file.
+    * Simulation information and settings `.solver.sharpy` file.
+    * The dynamic forces file `.dyn.h5` (when required).
+    * The linear input files `.lininput.h5` (when required).
+    * The ROM settings file `.rom.h5` (when required).
+    
+    See the [link TBD] for a detailed description on the contents of each of the files
+    
 
 2. The `h5` files contain data of the FEM, aerodynamics, dynamic conditions. They are later read by SHARPy.
 
-3. The `.solver.txt` file contains the settings for SHARPy and is the file that is parsed to SHARPy.
+3. The `.solver.sharpy` file contains the settings for SHARPy and is the file that is parsed to SHARPy.
 
 __To run a SHARPy case__
 
@@ -255,7 +266,7 @@ SHARPy cases are therefore usually ran in the following way:
 
 3. Run SHARPy (ensure the environment is activated)
     ```bash
-    run_sharpy case.solver.txt
+    run_sharpy case.solver.sharpy
     ```
 
 ### Output
@@ -265,62 +276,79 @@ By default, the output is located in the `output` folder.
 The contents of the folder will typically be a `beam` and `aero` folders, which contain the output data that can then be
 loaded in Paraview.
 
-#### Run a test case
+#### Running (and modifiying) a test case
 
-__*TODO* review Geradin case and update sharpy calls__
-
-__TUTORIAL OUT OF DATE__
-
-
-
-1.  This command generates the required files for running a static, clamped beam: 
+1.  This command generates the required files for running a static, clamped beam case that is used as part of code 
+verification: 
     ```sh
     cd ../sharpy
     python ./tests/beam/static/geradin_cardona/generate_geradin.py
     ```
     
-Now you should see a success message, and if you check the
-`./tests/beam/static/geradin_cardona/` folder, you should see two new files:
-+ geradin_cardona.solver.txt
-+ geradin_cardona.fem.h5
+    The script will run and if you check the
+    `./tests/beam/static/geradin_cardona/` folder, you should see two new files:
+    + `geradin.solver.txt`
+    + `geradin.fem.h5`
 
-Try to open the `solver.txt` file and have a quick look. The `solver` file is
-the main settings file. We'll get deeper into this later.
+    Try to open the `solver.txt` file and have a quick look. The `solver` file is
+    the main settings file. We'll get deeper into this in a separate chapter.
 
-If you try to open the `fem.h5` file, you'll get an error or something meaningless. This is because the structural data is stored in [HDF5](https://support.hdfgroup.org/HDF5/) format, which is compressed binary.
+    If you try to open the `fem.h5` file, you'll get an error or something meaningless. This is because the structural data
+    is stored in [HDF5](https://support.hdfgroup.org/HDF5/) format, which is compressed binary.
 
-5. Now run it
+5. Run it (part 1)
 
[…]
