### 2019-11-12 afb10a0
Added preliminary Dockerfile for distribution
