+++ docs/source/content/casefiles.md
@@ -1,81 +1,0 @@
-# The SHARPy Case files
-
-##  The FEM file
-
-The case.fem.h5 file has several components. We go one by one:
-
-   - `num_node_elem [int]`: number of nodes per element.
-   
-        Always 3 in our case (3 nodes per structural elements - quadratic beam elements).
-    
-   - `num_elem [int]`: number of structural elements.
-    
-   - `num_node [int]`: number of nodes. 
-   
-       For simple structures, it is `num_elem*(num_node_elem - 1) - 1`. 
-       For more complicated ones, you need to calculate it properly.
-
-   - `coordinates [num_node, 3]` coordinates of the nodes in body-attached FoR.
-
-   - `connectivites [num_elem, num_node_elem]`: Beam element's connectivities. 
-   
-       Every row refers to an element, and the three integers in that row are the indices of the three nodes 
-       belonging to that elem. Now, the catch: the ordering is not as you'd think. Order them as `[0, 2, 1]`. 
-       That means, first one, last one, central one. The following image shows the node indices inside the 
-       circles representing the nodes, the element indices in blue and the resulting connectivities matrix next to it. 
-       Connectivities are tricky when considering complex configurations. Pay attention at the beginning and you'll 
-       save yourself a lot of trouble.
-
-
-```
-stiffness_db [:, 6, 6] database of stiffness matrices. The first dimension has as many elements as different stiffness matrices are in the model.
-
-elem_stiffness [num_elem] array of indices (starting at 0). Basically, it links every element (index) to the stiffness matrix index in stiffness_db. For example elem_stiffness[0] = 0; elem_stiffness[2] = 1 means that the element 0 has a stiffness matrix equal to stiffness_db[0, :, :], and the second element has a stiffness matrix equal to stiffness_db[1, :, :].
-
-The shape of a stiffness matrix, $\mathrm{S}$ is:
-S=⎡⎣⎢⎢⎢⎢⎢⎢⎢⎢⎢EAGAyGAzGJEIyEIz⎤⎦⎥⎥⎥⎥⎥⎥⎥⎥⎥
- 
-with the cross terms added if needed.
-
-mass_db and elem_mass follow the same scheme than the stiffness, but the mass matrix is given by:
-M=[mIξ~cgm−ξ~cgmJ]
- 
-where $m$ is the distributed mass per unit length [kg/m], $\tilde{\bullet}$ is the skew-symmetric matrix of a vector and $\xi_{cg}$ is the location of the centre of gravity with respect to the elastic axis in MATERIAL (local) FoR.
-And what is the Material FoR? This is an important point, because all the inputs that move WITH the beam are in material FoR. For example: follower forces, stiffness, mass, lumped masses...
-
-
-
-The material frame of reference is noted as $B$. Essentially, the $x$ component is tangent to the beam in the increasing node ordering, $z$ looks up generally and $y$ is oriented such that the FoR is right handed.
-
-In the practice (vertical surfaces, structural twist effects...) it is more complicated than this. The only sure thing about $B$ is that its $x$ direction is tangent to the beam in the increasing node number direction. However, with just this, we have an infinite number of potential reference frames, with $y$ and $z$ being normal to $x$ but rotating around it. The solution is to indicate a for_delta, or frame of reference delta vector ($\Delta$).
-
-
-
-Now we can define unequivocally the material frame of reference. With $x_B$ and $\Delta$ defining a plane, $y_b$ is chosen such that the $z$ component is oriented upwards with respect to the lifting surface.
-
-From this definition comes the only constraint to $\Delta$: it cannot be parallel to $x_B$.
-
[…]
