### 2019-11-12 e709754
Minimally functioning Docker image
