### 2019-11-12 d3ef7dd
Initial work on jupyter examples
