+++ docs/source/content/examples.rst
@@ -1,0 +1,16 @@
+Examples
+========
+
+A set of SHARPy examples created with Jupyter Notebooks is provided for users to interact and modify
+cases running on SHARPy.
+
+SHARPy Examples:
+
+.. toctree::
+    example_notebooks/linear_goland_flutter
+
+Download links:
+
+* :download:`./example_notebooks/linear_goland_flutter.ipynb`
+
+
