### 2019-11-13 d854aff
Goland flutter analysis example
