+++ docs/source/content/examples.rst
@@ -4,12 +4,14 @@ Examples
 A set of SHARPy examples created with Jupyter Notebooks is provided for users to interact and modify
 cases running on SHARPy.
 
-SHARPy Examples:
-
 .. toctree::
+    :maxdepth: 1
+    :caption: SHARPy Examples
+
     example_notebooks/linear_goland_flutter
 
-Download links:
+Downloadable files
+------------------
 
 * :download:`./example_notebooks/linear_goland_flutter.ipynb`
 
