### 2019-11-13 26cb811
Added nbsphinx to env_linux
