+++ readthedocs.yml
@@ -1,0 +1,2 @@
+conda:
+  file: utils/environment_linux.yml
