+++ readthedocs.yml
@@ -1,2 +1,2 @@
 conda:
-  file: utils/environment_linux.yml
+  - nbsphinx
