### 2019-11-13 c400036
Fixed krylov rom tests that used random matrices each time
