### 2019-11-13 50ece98
Removed linear warning + fixed files appearing in generate cases tests
