### 2019-11-13 e7882a6
Fixed bug with gustvelocity field
