### 2019-11-13 9c62ea7
Added simple_HALE example to docs
