+++ docs/source/conf.py
@@ -81,7 +81,7 @@ author = 'Aeroelastics Lab, Imperial College London'
 # The short X.Y version.
 version = '1.0'
 # The full version, including alpha/beta/rc tags.
-release = '1.0.0-rc1'
+release = '1.0.0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
