### 2019-11-13 0339e0d
More generic nbsphinx requirement
