### 2019-11-14 b11aab1
Fixed list spacing in jupyter notebooks. Lists req blank line in between
