### 2019-11-14 0c1595a
Dockerfile makes a working container
