### 2019-11-14 57e70c4
Dockerfile generates smaller image now
