### 2019-11-14 ffd457a
Removed matplotlib dependency of core SHARPy
