### 2019-11-14 9b2bfee
Added conve environment for lightweight distribution
