+++ utils/environment_minimal.yml
@@ -1,0 +1,62 @@
+name: sharpy_minimal
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1=main
+  - blas=1.0=mkl
+  - ca-certificates=2019.10.16=0
+  - certifi=2019.9.11=py37_0
+  - colorama=0.4.1=py37_0
+  - configobj=5.0.6=py37_1
+  - dill=0.3.1.1=py37_0
+  - eigen=3.3.7=hfd86e86_0
+  - h5py=2.9.0=py37h7918eee_0
+  - hdf5=1.10.4=hb1b8bf9_0
+  - icu=58.2=h9c2bf20_1
+  - intel-openmp=2019.4=243
+  - lapack=3.6.1=ha44fe06_2
+  - libedit=3.1.20181209=hc058e9b_0
+  - libffi=3.2.1=hd88cf55_4
+  - libgcc-ng=9.1.0=hdf63c60_0
+  - libgfortran=3.0.0=1
+  - libgfortran-ng=7.3.0=hdf63c60_0
+  - libstdcxx-ng=9.1.0=hdf63c60_0
+  - libxml2=2.9.9=hea5a465_1
+  - libxslt=1.1.33=h7d1a2b0_0
+  - lxml=4.4.1=py37hefd8a0e_0
+  - mkl=2019.4=243
+  - mkl-service=2.3.0=py37he904b0f_0
+  - mkl_fft=1.0.15=py37ha843d7b_0
+  - mkl_random=1.1.0=py37hd6b4f25_0
+  - ncurses=6.1=he6710b0_1
+  - numpy=1.17.3=py37hd14ec0e_0
+  - numpy-base=1.17.3=py37hde5b4d6_0
+  - openssl=1.1.1d=h7b6447c_3
+  - pandas=0.25.2=py37he6710b0_0
+  - pip=19.3.1=py37_0
+  - python=3.7.5=h0371630_0
+  - python-dateutil=2.8.1=py_0
+  - pytz=2019.3=py_0
+  - pyyaml=5.1.2=py37h7b6447c_0
+  - readline=7.0=h7b6447c_5
+  - scipy=1.3.1=py37h7c811a0_0
+  - setuptools=41.6.0=py37_0
+  - six=1.13.0=py37_0
+  - sqlite=3.30.1=h7b6447c_0
+  - tk=8.6.8=hbc83047_0
+  - wheel=0.33.6=py37_0
+  - xlrd=1.2.0=py37_0
+  - xz=5.2.4=h14c3975_4
+  - yaml=0.1.7=had09818_2
+  - zlib=1.2.11=h7b6447c_3
+  - pip:
+    - apptools==4.5.0
+    - envisage==4.8.0
+    - mayavi==4.7.1
+    - pyface==6.1.2
+    - pygments==2.4.2
+    - traits==5.1.2
+    - traitsui==6.1.3
+    - vtk==8.1.2
+
