### 2019-11-14 5130b2f
Updated Dockerfile and bashrc to use sharpy_minimal environment
