### 2019-11-14 b8cb907
Added tmux and vim to the docker for convenience
