+++ Dockerfile
@@ -7,7 +7,7 @@ ENV PATH=${PATH}:/miniconda3/bin
 
 # Development tools including compilers
 RUN yum groupinstall "Development Tools" -y && \
-    yum install -y mesa-libGL libXt libXt-devel wget gcc-gfortran lapack && \
+    yum install -y mesa-libGL libXt libXt-devel wget gcc-gfortran lapack vim-minimal tmux && \
     yum clean all
 
 # Install miniconda
