### 2019-11-14 adaf143
Forgot to add mkl-includes to the enviroment for UVLM compilation
