### 2019-11-14 04b7a3e
Error if setting type unknown plus fix settings no ctype option for bool
