+++ sharpy/generators/gustvelocityfield.py
@@ -1,347 +1,8 @@
 import numpy as np
-# from abc import ABCMeta, abstractmethod
-from abc import ABCMeta
 
 import sharpy.utils.generator_interface as generator_interface
 import sharpy.utils.settings as settings
 import sharpy.utils.exceptions as exc
-
-dict_of_gusts = {}
-
-
-def gust(arg):
-    global dict_of_gusts
-    try:
-        arg._gust_id
-    except:
-        raise AttributeError('Class defined as gust has no _gust_id attribute')
-    dict_of_gusts[arg._gust_id] = arg
-    return arg
-
-
-#@gust
-class BaseGust(metaclass=ABCMeta):
-    pass
-# class BaseGust():
-#     _gust_id = 'BaseGust'
-#
-#     settings_types = dict()
-#     settings_default = dict()
-#     settings_description = dict()
-#
-#     def gust_shape(x, y, z, time=0):
-#         return np.zeros((3))
-
-
-# class one-cos(BaseGust):
-@gust
-class one_minus_cos(BaseGust):
-    r"""
-        Discrete gust model
-        .. math:: U_z = \frac{u_{de}}{2}\left[1-\cos\left(\frac{2\pi x}{S}\right)\right]
-    """
-    _gust_id = '1-cos'
-
-    settings_types = dict()
-    settings_default = dict()
-    settings_description = dict()
-
-    settings_types['gust_length'] = 'float'
-    settings_default['gust_length'] = 0.0
-    settings_description['gust_length'] = 'Length of gust'
-
-    settings_types['gust_intensity'] = 'float'
-    settings_default['gust_intensity'] = 0.0
-    settings_description['gust_intensity'] = 'Intensity of the gust'
-
-    setting_table = settings.SettingsTable()
-    __doc__ += setting_table.generate(settings_types, settings_default, settings_description)
-
-    def __init__(self):
-        pass
-
-    def initialise(self, in_dict):
-        self.in_dict = in_dict
-        settings.to_custom_types(self.in_dict, self.settings_types, self.settings_default)
-        self.settings = self.in_dict
-
-    def gust_shape(self, x, y, z, time=0):
-        gust_length = self.settings['gust_length'].value
-        gust_intensity = self.settings['gust_intensity'].value
-
-        vel = np.zeros((3,))
-        if x > 0.0 or x < -gust_length:
-            return vel
-
-        vel[2] = (1.0 - np.cos(2.0*np.pi*x/gust_length))*gust_intensity*0.5
-        return vel
-
-
-@gust
-class DARPA(BaseGust):
-    r"""
-    Discrete, non-uniform span model
-    .. math:: U_z = \frac{u_{de}}{2}\left[1-\cos\left(\frac{2\pi x}{S}\right)\right]\cos\left(\frac{\pi y}{b}\right)
-    """
-    _gust_id = 'DARPA'
-
-    settings_types = dict()
-    settings_default = dict()
-    settings_description = dict()
-
-    settings_types['gust_length'] = 'float'
-    settings_default['gust_length'] = 0.0
-    settings_description['gust_length'] = 'Length of gust'
-
-    settings_types['gust_intensity'] = 'float'
-    settings_default['gust_intensity'] = 0.0
-    settings_description['gust_intensity'] = 'Intensity of the gust'
-
-    settings_types['span'] = 'float'
-    settings_default['span'] = 0.
-    settings_description['span'] = 'Wing span'
-
-    setting_table = settings.SettingsTable()
-    __doc__ += setting_table.generate(settings_types, settings_default, settings_description)
-
-    def initialise(self, in_dict):
-        self.in_dict = in_dict
[…]
