### 2019-11-14 8ce797a
GustVel generator init classes and u_inf properties
