+++ sharpy/generators/gustvelocityfield.py
@@ -8,6 +8,9 @@ import sharpy.utils.exceptions as exc
 
 dict_of_gusts = {}
 
+doc_settings_description = 'The ``GustVelocityField`` generator takes the following settings as a dictionary assigned' \
+                           'to ``gust_parameters``.'
+
 
 def gust(arg):
     global dict_of_gusts
@@ -21,7 +24,28 @@ def gust(arg):
 
 #@gust
 class BaseGust(metaclass=ABCMeta):
-    pass
+
+    def __init__(self):
+        self.settings = dict()
+        self._u_inf = None
+        self._u_inf_direction = None
+
+    @property
+    def u_inf(self):
+        return self._u_inf
+
+    @u_inf.setter
+    def u_inf(self, value):
+        self._u_inf = value
+
+    @property
+    def u_inf_direction(self):
+        return self._u_inf_direction
+
+    @u_inf_direction.setter
+    def u_inf_direction(self, value):
+        self._u_inf_direction = value
+
 # class BaseGust():
 #     _gust_id = 'BaseGust'
 #
@@ -37,8 +61,10 @@ class BaseGust(metaclass=ABCMeta):
 @gust
 class one_minus_cos(BaseGust):
     r"""
-        Discrete gust model
+    One minus cos gust (single bump)
         .. math:: U_z = \frac{u_{de}}{2}\left[1-\cos\left(\frac{2\pi x}{S}\right)\right]
+
+    This gust can be used by using the setting ``gust_shape = 1-cos`` in ``GustVelocityField``.
     """
     _gust_id = '1-cos'
 
@@ -55,15 +81,11 @@ class one_minus_cos(BaseGust):
     settings_description['gust_intensity'] = 'Intensity of the gust'
 
     setting_table = settings.SettingsTable()
-    __doc__ += setting_table.generate(settings_types, settings_default, settings_description)
-
-    def __init__(self):
-        self.settings = None
+    __doc__ += setting_table.generate(settings_types, settings_default, settings_description, doc_settings_description)
 
     def initialise(self, in_dict):
-        self.in_dict = in_dict
-        settings.to_custom_types(self.in_dict, self.settings_types, self.settings_default)
-        self.settings = self.in_dict
+        self.settings = in_dict
+        settings.to_custom_types(self.settings, self.settings_types, self.settings_default)
 
     def gust_shape(self, x, y, z, time=0):
         gust_length = self.settings['gust_length'].value
@@ -81,7 +103,10 @@ class one_minus_cos(BaseGust):
 class DARPA(BaseGust):
     r"""
     Discrete, non-uniform span model
+
     .. math:: U_z = \frac{u_{de}}{2}\left[1-\cos\left(\frac{2\pi x}{S}\right)\right]\cos\left(\frac{\pi y}{b}\right)
+
+    This gust can be used by using the setting ``gust_shape = DARPA`` in ``GustVelocityField``.
     """
     _gust_id = 'DARPA'
 
@@ -102,12 +127,11 @@ class DARPA(BaseGust):
     settings_description['span'] = 'Wing span'
 
     setting_table = settings.SettingsTable()
-    __doc__ += setting_table.generate(settings_types, settings_default, settings_description)
+    __doc__ += setting_table.generate(settings_types, settings_default, settings_description, doc_settings_description)
 
     def initialise(self, in_dict):
-        self.in_dict = in_dict
-        settings.to_custom_types(self.in_dict, self.settings_types, self.settings_default)
-        self.settings = self.in_dict
+        self.settings = in_dict
+        settings.to_custom_types(self.settings, self.settings_types, self.settings_default)
 
     def gust_shape(self, x, y, z, time=0):
[…]
