### 2019-11-14 fdea8d5
Minor doc changes to GustGenerator
