+++ sharpy/generators/gustvelocityfield.py
@@ -1,10 +1,8 @@
 import numpy as np
-# from abc import ABCMeta, abstractmethod
 from abc import ABCMeta
 
 import sharpy.utils.generator_interface as generator_interface
 import sharpy.utils.settings as settings
-import sharpy.utils.exceptions as exc
 
 dict_of_gusts = {}
 
@@ -15,15 +13,19 @@ doc_settings_description = 'The ``GustVelocityField`` generator takes the follow
 def gust(arg):
     global dict_of_gusts
     try:
-        arg._gust_id
-    except:
-        raise AttributeError('Class defined as gust has no _gust_id attribute')
-    dict_of_gusts[arg._gust_id] = arg
+        arg.gust_id
+    except AttributeError:
+        raise AttributeError('Class defined as gust has no gust_id attribute')
+    dict_of_gusts[arg.gust_id] = arg
     return arg
 
 
-#@gust
+# @gust
 class BaseGust(metaclass=ABCMeta):
+
+    settings_types = dict()
+    settings_default = dict()
+    settings_description = dict()
 
     def __init__(self):
         self.settings = dict()
@@ -46,8 +48,9 @@ class BaseGust(metaclass=ABCMeta):
     def u_inf_direction(self, value):
         self._u_inf_direction = value
 
+
 # class BaseGust():
-#     _gust_id = 'BaseGust'
+#     gust_id = 'BaseGust'
 #
 #     settings_types = dict()
 #     settings_default = dict()
@@ -64,9 +67,9 @@ class one_minus_cos(BaseGust):
     One minus cos gust (single bump)
         .. math:: U_z = \frac{u_{de}}{2}\left[1-\cos\left(\frac{2\pi x}{S}\right)\right]
 
-    This gust can be used by using the setting ``gust_shape = 1-cos`` in ``GustVelocityField``.
+    This gust can be used by using the setting ``gust_shape = '1-cos'`` in ``GustVelocityField``.
     """
-    _gust_id = '1-cos'
+    gust_id = '1-cos'
 
     settings_types = dict()
     settings_default = dict()
@@ -95,7 +98,7 @@ class one_minus_cos(BaseGust):
         if x > 0.0 or x < -gust_length:
             return vel
 
-        vel[2] = (1.0 - np.cos(2.0*np.pi*x/gust_length))*gust_intensity*0.5
+        vel[2] = (1.0 - np.cos(2.0 * np.pi * x / gust_length)) * gust_intensity * 0.5
         return vel
 
 
@@ -106,9 +109,9 @@ class DARPA(BaseGust):
 
     .. math:: U_z = \frac{u_{de}}{2}\left[1-\cos\left(\frac{2\pi x}{S}\right)\right]\cos\left(\frac{\pi y}{b}\right)
 
-    This gust can be used by using the setting ``gust_shape = DARPA`` in ``GustVelocityField``.
+    This gust can be used by using the setting ``gust_shape = 'DARPA'`` in ``GustVelocityField``.
     """
-    _gust_id = 'DARPA'
+    gust_id = 'DARPA'
 
     settings_types = dict()
     settings_default = dict()
@@ -142,8 +145,8 @@ class DARPA(BaseGust):
         if x > 0.0 or x < -gust_length:
             return vel
 
-        vel[2] = (1.0 - np.cos(2.0*np.pi*x/gust_length))*gust_intensity*0.5
-        vel[2] *= -np.cos(y/span*np.pi)
+        vel[2] = (1.0 - np.cos(2.0 * np.pi * x / gust_length)) * gust_intensity * 0.5
+        vel[2] *= -np.cos(y / span * np.pi)
         return vel
 
 
[…]
