### 2019-11-15 9bebd32
Finalised new gust velocity field and associated docs
