+++ docs/source/includes/generators/dynamiccontrolsurface/DynamicControlSurface.rst
@@ -1,7 +1,5 @@
 DynamicControlSurface
 ---------------------
 
-
-
 .. autoclass:: sharpy.generators.dynamiccontrolsurface.DynamicControlSurface
 	:members:
