+++ docs/source/includes/generators/turbvelocityfieldbts/TurbVelocityFieldBts.rst
@@ -1,7 +1,5 @@
 TurbVelocityFieldBts
 --------------------
 
-
-
 .. autoclass:: sharpy.generators.turbvelocityfieldbts.TurbVelocityFieldBts
 	:members:
