+++ docs/source/includes/generators/bumpvelocityfield/BumpVelocityField.rst
@@ -1,7 +1,5 @@
 BumpVelocityField
 -----------------
 
-
-
 .. autoclass:: sharpy.generators.bumpvelocityfield.BumpVelocityField
 	:members:
