+++ docs/source/includes/generators/index.rst
@@ -1,5 +1,6 @@
-Field Generators
-----------------
+Generators
+----------
+
 
 Velocity field generators prescribe the flow conditions for your problem. For instance, you can have an aircraft at
 a prescribed fixed location in a velocity field towards the aircraft. Alternatively, you can have a free moving
@@ -8,7 +9,14 @@ aircraft in a static velocity field.
 Dynamic Control Surface generators enable the user to prescribe a certain control surface deflection in time.
 
 .. toctree::
-    :glob:
+	:maxdepth: 1
 
-    ../includes/generators/*
-
+	./bumpvelocityfield/index
+	./dynamiccontrolsurface/index
+	./gridbox/index
+	./gustvelocityfield/index
+	./shearvelocityfield/index
+	./steadyvelocityfield/index
+	./trajectorygenerator/index
+	./turbvelocityfield/index
+	./turbvelocityfieldbts/index
