+++ docs/source/includes/generators/gustvelocityfield/GustVelocityField.rst
@@ -1,7 +1,5 @@
 GustVelocityField
 -----------------
 
-
-
 .. autoclass:: sharpy.generators.gustvelocityfield.GustVelocityField
 	:members:
