+++ docs/source/includes/generators/gridbox/GridBox.rst
@@ -1,7 +1,5 @@
 GridBox
 -------
 
-
-
 .. autoclass:: sharpy.generators.gridbox.GridBox
 	:members:
