### 2019-11-15 604b9e5
removed commented lines in gustvelocityfield
