### 2019-11-15 d4ee345
updated function calls for numpy 1.17
