+++ sharpy/utils/generate_cases.py
@@ -131,7 +131,7 @@ def read_column_sheet_type01(excel_file_name, excel_sheet, column_name):
 
     xls = pd.ExcelFile(excel_file_name)
     excel_db = pd.read_excel(xls, sheet_name=excel_sheet)
-    num_elem = excel_db.index._stop - 2
+    num_elem = excel_db.index.stop - 2
 
     if excel_db[column_name][1] == 'one_int':
         var = excel_db[column_name][2]
@@ -149,7 +149,7 @@ def read_column_sheet_type01(excel_file_name, excel_sheet, column_name):
         print("ERROR: not recognized number type")
 
     if 'vec' in excel_db[column_name][1]:
-        for i in range(2, excel_db.index._stop):
+        for i in range(2, excel_db.index.stop):
             var[i - 2] = excel_db[column_name][i]
 
     return var
