### 2019-11-15 452b7f5
Updated environment for linux with numpy and doc generation packages
