+++ utils/environment_linux.yml
@@ -5,165 +5,129 @@ channels:
 dependencies:
   - _libgcc_mutex=0.1=main
   - alabaster=0.7.12=py37_0
-  - asn1crypto=0.24.0=py37_0
-  - babel=2.6.0=py37_0
-  - backcall=0.1.0=py37_0
-  - basemap=1.2.0=py37h705c2d8_0
+  - asn1crypto=1.2.0=py37_0
+  - attrs=19.3.0=py_0
+  - babel=2.7.0=py_0
   - blas=1.0=mkl
-  - bleach=3.1.0=py37_0
-  - bzip2=1.0.6=h14c3975_5
-  - ca-certificates=2019.5.15=1
-  - certifi=2019.6.16=py37_1
-  - cffi=1.11.5=py37he75722e_1
-  - chardet=3.0.4=py37_1
+  - bleach=3.1.0=py_0
+  - ca-certificates=2019.9.11=hecc5488_0
+  - certifi=2019.9.11=py37_0
+  - cffi=1.13.1=py37h2e261b9_0
+  - chardet=3.0.4=py37_1003
   - colorama=0.4.1=py37_0
-  - commonmark=0.8.1=py_0
+  - commonmark=0.9.0=py_0
   - configobj=5.0.6=py37_1
-  - coverage=4.5.3=py37h7b6447c_0
-  - cryptography=2.5=py37h1ba5d50_0
-  - curl=7.63.0=hbc83047_1000
+  - cryptography=2.8=py37h1ba5d50_0
   - cycler=0.10.0=py37_0
-  - cython=0.29.5=py37he6710b0_0
-  - dbus=1.13.6=h746ee38_0
-  - decorator=4.3.2=py37_0
-  - dill=0.3.0=py37_0
-  - docutils=0.14=py37_0
+  - dbus=1.13.12=h746ee38_0
+  - decorator=4.4.1=py_0
+  - defusedxml=0.6.0=py_0
+  - dill=0.3.1.1=py37_0
+  - docutils=0.15.2=py37_0
   - eigen=3.3.7=hfd86e86_0
-  - entrypoints=0.3=py37_0
+  - entrypoints=0.3=py37_1000
   - expat=2.2.6=he6710b0_0
   - fontconfig=2.13.0=h9420a91_0
   - freetype=2.9.1=h8a8886c_1
-  - future=0.17.1=py37_0
-  - geos=3.6.2=heeff764_2
-  - glib=2.56.2=hd408876_0
-  - gmp=6.1.2=h6c8ec71_1
+  - future=0.18.2=py37_0
+  - glib=2.63.1=h5a9c865_0
   - gst-plugins-base=1.14.0=hbbd80ab_1
   - gstreamer=1.14.0=hb453b48_1
   - h5py=2.9.0=py37h7918eee_0
-  - hdf4=4.2.13=h3ca952b_2
   - hdf5=1.10.4=hb1b8bf9_0
   - icu=58.2=h9c2bf20_1
   - idna=2.8=py37_0
   - imagesize=1.1.0=py37_0
-  - intel-openmp=2019.1=144
-  - ipykernel=5.1.0=py37h39e3cac_0
-  - ipython=7.2.0=py37h39e3cac_0
-  - ipython_genutils=0.2.0=py37_0
-  - ipywidgets=7.4.2=py37_0
-  - jedi=0.13.2=py37_0
-  - jinja2=2.10=py37_0
+  - importlib_metadata=0.23=py37_0
+  - intel-openmp=2019.4=243
+  - ipython_genutils=0.2.0=py_1
[…]
