### 2019-11-15 a1110f9
Badges for readme and documentation
