### 2019-11-15 6e0f02e
Added docker integration
