+++ docs/source/content/installation.md
@@ -8,26 +8,92 @@ The following step by step tutorial will guide you through the installation proc
 __Operating System Requirements__
 
 SHARPy is being developed and tested on the following operating systems:
-+ CentOS 7
++ CentOS 7 and CentOS 8
 + Ubuntu 18.04 LTS
-+ MacOS Sierra
++ MacOS Mojave and Catalina
+
+It is also available to the vast majority of operating systems that are supported
+by Docker, including Windows!
 
 __Required Distributions__
 
-+ Anaconda Python 3.5.3
-+ GCC 5.0 or higher (recommended)
++ Anaconda Python 3.7
++ GCC 6.0 or higher (recommended)
 
 
 __GitHub Repositories__
 
 + [SHARPy](http://github.com/imperialcollegelongon/sharpy)
 
+SHARPy can be obtained from Docker Hub, or can be built from scratch.
+If what you want is to give it a go and run some static or simple dynamic cases,
+we recommend the Docker route. If you want to check the code, modify it and
+compile the libraries with custom flags, build it from source.
+
+## Using SHARPy from a Docker container
+
+Docker containers are similar to lightweight virtual machines. The SHARPy container
+distributed through [Docker Hub](https://hub.docker.com/) is a CentOS 8
+machine with the libraries compiled with `gfortran` and `g++` and an
+Anaconda Python distribution.
+
+Make sure your machine has Docker working. The instructions are here:
+[link](https://docs.docker.com/v17.09/engine/installation/).
+
+You might want to run a test in your terminal:
+```
+docker pull hello-world
+docker run hello-world
+```
+If this works, you're good to go!
+
+First, obtain the SHARPy docker container:
+```
+docker pull fonsocarre/sharpy:latest
+```
+
+Now you can run it:
+```
+docker run -it fonsocarre/sharpy:latest --name sharpy
+```
+You should see a welcome dialog such as:
+```
+>>>> docker run -it fonsocarre/sharpy:dev_docker
+SHARPy added to PATH from the directory: /sharpy/bin
+=======================================================================
+Welcome to the Docker image of SHARPy
+SHARPy is located in /sharpy and the
+environment is already set up!
+Copyright Imperial College London. Released under BSD 3-Clause license.
+=======================================================================
+SHARPy> 
+```
+You are now good to go.
+
+It is important to note that a docker container runs as an independant
+operating system with no access to your hard drive. If you want to copy your own
+files, run the container and from another terminal run:
+```
+docker cp my_file.txt sharpy:/my_file.txt     # copy from host to container
+docker cp sharpy:/my_file.txt my_file.txt     # copy from container to host
+```
+The `sharpy:` part is the `--name` argument you wrote in the `docker run` command.
+
+You can run the test suite once inside the container as:
+```
+cd sharpy
+python -m unittest
+```
+
+**Enjoy!**
+
+
+## Building SHARPy from source
+
 SHARPy depends on the UVLM and xbeam repositories that are also found on GitHub:
 
 + [xbeam](http://github.com/imperialcollegelondon/xbeam)
 + [UVLM](http://github.com/imperialcollegelondon/UVLM)
-
-## Installing SHARPy
 
 ### Set up the folder structure
 
@@ -236,11 +302,11 @@ You are now ready to run SHARPy cases from the terminal.
 SHARPy cases are usually structured in the following way:
 
 1. The `generate_case.py` file: contains the setup of the problem, like the geometry, flight conditions etc.
-This script creates two output files that will then be used by SHARPy, `.fem.h5` and the `.solver.txt` file.
[…]
