### 2019-11-15 460be60
Removed the pull from dev_docker in the Dockfile, now it is from master
