### 2019-11-15 407ec33
moved comments on CI and Zenodo to bottom of readme
