+++ README.md
@@ -1,40 +1,13 @@
 # Simulation of High Aspect Ratio aeroplanes in Python [SHARPy]
 
 ![Version badge](https://img.shields.io/endpoint.svg?url=https%3A%2F%2Fraw.githubusercontent.com%2FImperialCollegeLondon%2Fsharpy%2Fmaster%2F.version.json)
+[![Build Status](https://travis-ci.org/ImperialCollegeLondon/sharpy.svg?branch=master)](https://travis-ci.org/ImperialCollegeLondon/sharpy) 
+[![Documentation Status](https://readthedocs.org/projects/ic-sharpy/badge/?version=master)](https://ic-sharpy.readthedocs.io/en/master/?badge=master)
 [![codecov](https://codecov.io/gh/ImperialCollegeLondon/sharpy/branch/master/graph/badge.svg)](https://codecov.io/gh/ImperialCollegeLondon/sharpy)
 [![License](https://img.shields.io/badge/License-BSD%203--Clause-blue.svg)](https://opensource.org/licenses/BSD-3-Clause) 
-[![Documentation Status](https://readthedocs.org/projects/ic-sharpy/badge/?version=master)](https://ic-sharpy.readthedocs.io/en/master/?badge=master)
 [![status](https://joss.theoj.org/papers/f7ccd562160f1a54f64a81e90f5d9af9/status.svg)](https://joss.theoj.org/papers/f7ccd562160f1a54f64a81e90f5d9af9)
-
-__Zenodo__
-
-The _current_ version of SHARPy is archived in Zenodo. If you need to cite it, use this: 
-
-
 [![DOI](https://zenodo.org/badge/70235936.svg)](https://zenodo.org/badge/latestdoi/70235936)
 
-
-SHARPy has also a _concept_ DOI which represents all the versions of the given software package, i.e. the concept of 
-the software package and the ensemble of versions.
-
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3531965.svg)](https://doi.org/10.5281/zenodo.3531965)
-
-For more information on citing and Zenodo, [read more](https://help.zenodo.org/#versioning)
-
-__Continuous Integration Status__
-
-SHARPy uses Continuous Integration to control the integrity of its code. The status in the release and develop branches
-is:
-
-Master:
-
-[![Build Status](https://travis-ci.org/ImperialCollegeLondon/sharpy.svg?branch=master)](https://travis-ci.org/ImperialCollegeLondon/sharpy) 
-
-Develop:
-
-[![Build Status](https://travis-ci.org/ImperialCollegeLondon/sharpy.svg?branch=develop)](https://travis-ci.org/ImperialCollegeLondon/sharpy)
-
-## SHARPy
 
 SHARPy is an aeroelastic analysis package at the Department of Aeronautics, Imperial
 College London. It can be used for the structural, aerodynamic and aeroelastic analysis of flexible aircraft, flying
@@ -91,4 +64,30 @@ For the latest documentation, see the [installation docs](https://ic-sharpy.read
 
 If you think you can add some useful feature to SHARPy, by all means, check out
 the [collaboration guide](https://ic-sharpy.readthedocs.io/en/latest/content/contributing.html).
+
+## Citing SHARPy
+
+The _current_ version of SHARPy is archived in Zenodo. If you need to cite it, use this: 
+
+
+[![DOI](https://zenodo.org/badge/70235936.svg)](https://zenodo.org/badge/latestdoi/70235936)
+
+
+SHARPy has also a _concept_ DOI which represents all the versions of the given software package, i.e. the concept of 
+the software package and the ensemble of versions.
+
[…]
