### 2019-11-15 8c9f7e4
Updated MacOS environment
