+++ utils/environment_macos.yml
@@ -1,76 +1,118 @@
 name: sharpy_env
 channels:
+  - conda-forge
   - defaults
 dependencies:
+  - alabaster=0.7.12=py37_0
+  - asn1crypto=1.2.0=py37_0
+  - attrs=19.3.0=py_0
+  - babel=2.7.0=py_0
   - blas=1.0=mkl
-  - ca-certificates=2019.1.23=0
-  - certifi=2018.8.24=py35_1
-  - colorama=0.3.9=py35h1d66b2b_0
-  - cycler=0.10.0=py35hb89929e_0
-  - dbus=1.13.6=h90a0687_0
+  - bleach=3.1.0=py_0
+  - ca-certificates=2019.9.11=hecc5488_0
+  - certifi=2019.9.11=py37_0
+  - cffi=1.13.1=py37hb5b8e2f_0
+  - chardet=3.0.4=py37_1003
+  - colorama=0.4.1=py37_0
+  - commonmark=0.9.0=py_0
+  - cryptography=2.8=py37ha12b0ac_0
+  - cycler=0.10.0=py37_0
+  - decorator=4.4.1=py_0
+  - defusedxml=0.6.0=py_0
+  - dill=0.3.1.1=py37_0
+  - docutils=0.15.2=py37_0
   - eigen=3.3.7=h04f5b5a_0
-  - expat=2.2.6=h0a44026_0
-  - fontconfig=2.12.6=h9850fc5_0
-  - freetype=2.8=h12048fb_1
-  - future=0.16.0=py35_2
-  - gettext=0.19.8.1=h15daf44_3
-  - glib=2.56.2=hd9629dc_0
-  - h5py=2.7.1=py35he1c7800_0
-  - hdf5=1.10.1=ha036c08_1
+  - entrypoints=0.3=py37_1000
+  - freetype=2.9.1=hb4e5f40_0
+  - future=0.18.2=py37_0
+  - h5py=2.9.0=py37h3134771_0
+  - hdf5=1.10.4=hfa1e0ec_0
   - icu=58.2=h4b95b61_1
-  - intel-openmp=2019.3=199
-  - jpeg=9b=he5867d9_2
-  - jsoncpp=1.8.4=h04f5b5a_0
-  - kiwisolver=1.0.1=py35h219a9d8_0
+  - idna=2.8=py37_0
+  - imagesize=1.1.0=py37_0
+  - importlib_metadata=0.23=py37_0
+  - intel-openmp=2019.4=233
+  - ipython_genutils=0.2.0=py_1
+  - jinja2=2.10.3=py_0
+  - jsonschema=3.1.1=py37_0
+  - jupyter_core=4.5.0=py_0
+  - kiwisolver=1.1.0=py37h0a44026_0
+  - lapack=3.6.1=ha44fe06_2
   - libcxx=4.0.1=hcfea43d_1
   - libcxxabi=4.0.1=hcfea43d_1
   - libedit=3.1.20181209=hb402a30_0
   - libffi=3.2.1=h475c297_4
   - libgfortran=3.0.1=h93005f0_2
-  - libgfortran-ng=3.0.1=h93005f0_2
   - libiconv=1.15=hdd342a3_7
   - libpng=1.6.37=ha441bb4_0
-  - libtiff=4.0.10=hcb84e12_2
-  - libxml2=2.9.9=hab757c2_0
+  - libxml2=2.9.9=hf6e021a_1
   - libxslt=1.1.33=h33a18ac_0
-  - lxml=4.2.5=py35hef8c89e_0
-  - matplotlib=2.2.2=py35ha7267d0_0
-  - mkl=2018.0.3=1
[…]
