### 2019-11-15 7385fe4
Full stops in lists in readme
