### 2019-11-15 4b2639f
Added changelog before release of 1.0.1
