+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+[Full Changelog](https://github.com/imperialcollegelondon/sharpy/compare/1.0.0...HEAD)
+
+**Implemented enhancements:**
+
+- Stall Model [\#7](https://github.com/ImperialCollegeLondon/sharpy/issues/7)
+- New example of linearised flying wing [\#28](https://github.com/ImperialCollegeLondon/sharpy/pull/28) ([ngoiz](https://github.com/ngoiz))
+- SHARPy can now be obtained from a Docker Hub container [\#27](https://github.com/ImperialCollegeLondon/sharpy/pull/27) ([fonsocarre](https://github.com/fonsocarre))
+- Improved modal solver [\#26](https://github.com/ImperialCollegeLondon/sharpy/pull/26) ([fonsocarre](https://github.com/fonsocarre))
+- Updated function calls for latest numpy 1.17 [\#25](https://github.com/ImperialCollegeLondon/sharpy/pull/25) ([ngoiz](https://github.com/ngoiz))
+- Examples added to documentation [\#24](https://github.com/ImperialCollegeLondon/sharpy/pull/24) ([fonsocarre](https://github.com/fonsocarre))
+- Improved linear solver documentation and minor Krylov ROM fixes [\#23](https://github.com/ImperialCollegeLondon/sharpy/pull/23) ([ngoiz](https://github.com/ngoiz))
+- change log generator incorporated [\#22](https://github.com/ImperialCollegeLondon/sharpy/pull/22) ([ngoiz](https://github.com/ngoiz))
+
 
 
 \* *This Changelog was automatically generated by [github_changelog_generator](https://github.com/github-changelog-generator/github-changelog-generator)*
