### 2019-11-15 e5c4b3b
Updated version numbers before release
