### 2019-11-15 99482f3
Trying to fix CI bug with mayavi installation in headless servers
