### 2019-11-15 9028dac
Added vtk to list of conda-installed pakages
