+++ utils/environment_linux.yml
@@ -115,6 +115,7 @@ dependencies:
   - tornado=6.0.3=py37h7b6447c_0
   - traitlets=4.3.3=py37_0
   - urllib3=1.24.2=py37_0
+  - vtk=8.2
   - webencodings=0.5.1=py_1
   - wheel=0.33.6=py37_0
   - xlrd=1.2.0=py37_0
