### 2019-11-15 8aaaac1
Updated CI tests
