### 2019-11-15 8966c31
Updating travis file
