+++ .travis.yml
@@ -12,7 +12,6 @@ addons:
       - gfortran-7
 
 before_install:
-    # Link gcc-6 and g++-6 to their standard commands
   - sudo unlink /usr/bin/gcc && sudo ln -s /usr/bin/gcc-7 /usr/bin/gcc
   - sudo unlink /usr/bin/g++ && sudo ln -s /usr/bin/g++-7 /usr/bin/g++
   - sudo ln -s /usr/bin/gfortran-7 /usr/bin/gfortran
