+++ .travis.yml
@@ -26,7 +26,7 @@ before_install:
   - hash -r
   - export QT_QPA_PLATFORM='offscreen'
   - conda env create -f utils/environment_linux.yml
-  - conda activate sharpy_env
+  - source activate sharpy_env
 
 install:
   - echo $(pwd)
