### 2019-11-15 b908fd5
Update environment_linux.yml
