### 2019-11-16 ea68025
Update CHANGELOG.md
