### 2019-11-17 6b74f17
Update CHANGELOG.md
