### 2019-11-19 35de04d
Removed sparse if statement from mgs orthogonalisation
