### 2019-11-19 1bbfa4d
Fixed issue with images in the case file docs + minor improvements
