+++ docs/source/conf.py
@@ -45,6 +45,7 @@ extensions = [
     # 'sphinx.ext.githubpages',
     'sphinx.ext.napoleon',
     'nbsphinx',
+    'sphinx.ext.autosectionlabel',
 ]
 
 # Add any paths that contain templates here, relative to this directory.
