### 2019-11-19 10f2e0e
Remove from excluded list of docs sharpy.linear
