+++ sharpy/postproc/beamloads.py
@@ -1,12 +1,6 @@
-import os
-
 import numpy as np
-from tvtk.api import tvtk, write_data
-
-import sharpy.utils.cout_utils as cout
 from sharpy.utils.solver_interface import solver, BaseSolver
 import sharpy.utils.settings as settings
-import sharpy.utils.algebra as algebra
 import sharpy.structure.utils.xbeamlib as xbeamlib
 
 
