+++ sharpy/rom/krylov.py
@@ -1,6 +1,6 @@
+"""Krylov-subspaces model order reduction techniques
+"""
 import numpy as np
-import scipy.sparse as scsp
-import sharpy.linear.src.libsparse as libsp
 import scipy.linalg as sclalg
 import sharpy.linear.src.libss as libss
 import time
@@ -9,6 +9,8 @@ import sharpy.utils.cout_utils as cout
 import sharpy.utils.rom_interface as rom_interface
 import sharpy.utils.h5utils as h5
 import sharpy.rom.utils.krylovutils as krylovutils
+import warnings as warn
+
 
 @rom_interface.rom
 class Krylov(rom_interface.BaseRom):
@@ -71,6 +73,11 @@ class Krylov(rom_interface.BaseRom):
     settings_default['r'] = 1
     settings_description['r'] = 'Moments to match at the interpolation points'
 
+    settings_types['single_side'] = 'str'
+    settings_default['single_side'] = ''
+    settings_description['single_side'] = 'Construct the rom using a single side - ``C`` for controllability or ' \
+                                          '``O`` for observability. Leave blank for both.'
+
     settings_types['tangent_input_file'] = 'str'
     settings_default['tangent_input_file'] = ''
     settings_description['tangent_input_file'] = 'Filepath to .h5 file containing tangent interpolation vectors'
@@ -125,11 +132,9 @@ class Krylov(rom_interface.BaseRom):
                                       'could be that is not yet implemented'
                                       % self.algorithm)
 
-        self.frequency = np.array(self.settings['frequency'], dtype=complex)
+        self.frequency = np.array(self.settings['frequency'])
         self.r = self.settings['r'].value
         self.restart_arnoldi = self.settings['restart_arnoldi'].value
-        print(self.frequency)
-        # breakpoint()
         try:
             self.nfreq = self.frequency.shape[0]
         except AttributeError:
@@ -159,12 +164,9 @@ class Krylov(rom_interface.BaseRom):
         """
         self.ss = ss
 
-        try:
-            if self.settings['print_info']:
-                cout.cout_wrap('Model Order Reduction in progress...')
-                self.print_header()
-        except ValueError:
-            pass
+        if self.settings['print_info']:
+            cout.cout_wrap('Model Order Reduction in progress...')
+            self.print_header()
 
         if self.ss.dt is None:
             self.sstype = 'ct'
@@ -183,19 +185,20 @@ class Krylov(rom_interface.BaseRom):
         if not self.stable:
             # pass
             # Under development
-            self.restart()
-            TL, TR = self.stable_realisation()
-            self.ssrom = libss.ss(TL.T.dot(Ar.dot(TR)), TL.T.dot(Br), Cr.dot(TR), self.ss.D, self.ss.dt)
-            self.stable = self.check_stability(restart_arnoldi=self.restart_arnoldi)
+            warn.warnings('Stabilisation under development')
+            # TL, TR = self.restart()
+            # Wtr, Vr = self.restart()
+            # TL, TR = self.stable_realisation()
+            # self.ssrom = libss.ss(TL.T.dot(Ar.dot(TR)), TL.T.dot(Br), Cr.dot(TR), self.ss.D, self.ss.dt)
[…]
