### 2019-11-20 9263f4d
A few useful outputs to asymptotic stability
