+++ sharpy/postproc/asymptoticstability.py
@@ -46,8 +46,8 @@ class AsymptoticStability(BaseSolver):
 
     settings_types['export_eigenvalues'] = 'bool'
     settings_default['export_eigenvalues'] = False
-    settings_description['export_eigenvalues'] = 'Save eigenvalues and eigenvectors to file. ' \
-                                                 'Details in :func:`AsymptoticStability.export_eigenvalues`'
+    settings_description['export_eigenvalues'] = 'Save eigenvalues and eigenvectors to file. '
+                                                 # 'Details in :func:`AsymptoticStability.export_eigenvalues`'
 
     settings_types['display_root_locus'] = 'bool'
     settings_default['display_root_locus'] = False
