### 2019-11-20 97b0314
Attempt to fix RTD after destruction
