### 2019-11-20 6aaac22
Fixed some RTD warnings
