### 2019-11-21 27bc249
minor changes in docs for steplinearuvlm
