### 2019-11-21 3d28ecd
Fixes #32 - numpy arrays removed from default settings values
