### 2019-11-21 0a3e799
Revert "Improvements in documentation "
