### 2019-11-23 6be8f68
Added a short explanation on how to create the SHARPy settings
