### 2019-11-23 a48c0d3
Added the documentation for PreSharpy and changed settings to new format
