### 2019-11-24 cb5e4d0
Added settings_options dictionary to check against valid settings
