### 2019-11-24 d321b9b
Print available settings' options to documentation
