### 2019-11-24 51e4343
Fixed bug that appeared when settings were copied
