### 2019-11-24 a9ee27f
Change position of --name argument in docs
