### 2019-11-24 ImperialCollegeLondon/sharpy#36 @petebachant
Compare to other competing software in JOSS paper
From https://github.com/openjournals/joss-reviews/issues/1885.

>**State of the field:** Do the authors describe how this software compares to other commonly-used packages?
