### 2019-11-24 ImperialCollegeLondon/sharpy#36 @fonsocarre
@rafapalacios Which software packages would you include here as directly related? I can think of ZAERO, UM/NAST (none of them Open Source), and I guess OpenFast for WT aeroelasticity. Any input?
