### 2019-11-24 ImperialCollegeLondon/sharpy#37 @petebachant
run_theo_freq.py fails in Docker container
For https://github.com/openjournals/joss-reviews/issues/1885.

I'm assuming one would want to be able to run the example cases from the Docker container.

Firstly, I had to install matplotlib, since it's imported in the script. After that, I get the error:

```
Traceback (most recent call last):
  File "run_theo_freq.py", line 236, in <module>
    Sol = lin_aeroelastic.LinAeroEla(data0)
  File "/sharpy/sharpy/linear/src/lin_aeroelastic.py", line 118, in __init__
    self.lingebm_str = lingebm.FlexDynamic(self.tsstr, structure, beam_settings)
  File "/sharpy/sharpy/linear/src/lingebm.py", line 119, in __init__
    self.Nmodes = self.settings['num_modes']
KeyError: 'num_modes'
```
