### 2019-11-24 ImperialCollegeLondon/sharpy#37 @fonsocarre
Thanks @petebachant, you are 100% right. This script is old and somehow it survived the clean up. @ngoiz will check it, as the linear part is his stuff.

Concerning the matplotlib dependency. It is 50/50 in purpose. I decided when releasing v1.0.0 to remove matplotlib as a required dependency, specially for the docker image, to keep it light and compatible with headless machines. That's why the `utils/environment_linux.yml` has it, but `enviroment_minimal.yml` (the one for the Docker build) does not have it. Unfortunately, as I said, this script slipped through the cracks. My advice to @ngoiz is to wrap the plotting bits in a `try except` block (including the import) and use matplotlib if available, if not, print a warning and keep running.
