### 2019-11-24 ImperialCollegeLondon/sharpy#36 @rafapalacios
For aircraft application, ASWING (http://web.mit.edu/drela/Public/web/aswing/) is the only one with some online support, although you need to license it from MIT. ZAERO is an aerodynamic tool and it is full proprietary; and UM/NAST is a research code only referred to in papers.
For wind turbines, OpenFAST (https://github.com/openfast) and HAWC2 http://www.hawc2.dk/, cover a similar space, although they are both built using simple aerodynamics.
