### 2019-11-24 ImperialCollegeLondon/sharpy#37 @petebachant
That sounds reasonable to me!
