### 2019-11-24 ImperialCollegeLondon/sharpy#36 @fonsocarre
Thanks @rafapalacios, more or less what I thought. I'll write a paragraph about it and give you a shout when it is done in the pull request.
