### 2019-11-24 ImperialCollegeLondon/sharpy#37 @ngoiz
Thanks, Pete, for going through this. Indeed this is a legacy test case that slipped through the cracks and made it onto the release build. So the first step is that it will be removed from the release build and then I will incorporate it as a proper test case (or even tutorial since it is a nice example). PR coming soon for this.

And I agree on the matplotlib issues, those will be fixed as well with the try except option.
