### 2019-11-25 072dfa2
Modified settings check to account for ctypes as well
