+++ docs/source/includes/utils/settings/check_settings.rst
@@ -1,0 +1,4 @@
+check_settings
+--------------
+
+.. automodule:: sharpy.utils.settings.check_settings
