### 2019-11-25 ec1d5f2
Implemented options to a few solvers
