### 2019-11-25 c84132e
Fix bug calling warnings
