### 2019-11-25 b1991c5
Try except matplotlib in freqresponse post processor
