### 2019-11-25 e1847aa
Introduces custom exception for invalid settings
