+++ sharpy/utils/settings.py
@@ -207,7 +207,7 @@ def check_settings(settings, settings_types, settings_options):
         settings_options (dict): Dictionary of options (may be empty)
 
     Raises:
-        ValueError: if the setting is not allowed.
+        exception.NotValidSetting: if the setting is not allowed.
     """
     for k in settings_options:
         if settings_types[k] == 'int':
@@ -216,21 +216,18 @@ def check_settings(settings, settings_types, settings_options):
             except AttributeError:
                 value = settings[k]
             if value not in settings_options[k]:
-                raise ValueError('The setting %s with value %s is not one of the available '
-                                 'options: %s' % (k, value, settings_options[k]))
+                raise exceptions.NotValidSetting(k, value, settings_options[k])
 
         elif settings_types[k] == 'str':
             value = settings[k]
             if value not in settings_options[k] and value:
                 # checks that the value is within the options and that it is not an empty string.
-                raise ValueError('The setting %s with value %s is not one of the available '
-                                 'options: %s' % (k, value, settings_options[k]))
+                raise exceptions.NotValidSetting(k, value, settings_options[k])
 
         elif settings_types[k] == 'list(str)':
             for item in settings[k]:
                 if item not in settings_options[k] and item:
-                    raise ValueError('The setting %s with item %s is not one of the available '
-                                     'options: %s' % (k, item, settings_options[k]))
+                    raise exceptions.NotValidSetting(k, item, settings_options[k])
 
         else:
             pass  # no other checks implemented / required
