### 2019-11-25 a07bbfd
Added information about competing software in JOSS paper
