### 2019-11-25 9f6f855
Removed all matplotlib imports from the code and/or wrapped in try, expt
