+++ sharpy/solvers/nonlinearstaticmultibody.py
@@ -1,21 +1,11 @@
 import ctypes as ct
 import numpy as np
 
-# from sharpy.utils.settings import str2bool
-# import sharpy.utils.solver_interface as solver_interface
 from sharpy.utils.solver_interface import solver, BaseSolver, solver_from_string
 import sharpy.utils.settings as settings
-# import sharpy.utils.cout_utils as cout
-#
-# import scipy.linalg
 import sharpy.structure.utils.xbeamlib as xbeamlib
-# import sharpy.utils.algebra as algebra
-# import h5py as h5
-# import sharpy.utils.h5utils as h5utils
 import sharpy.utils.multibody as mb
-# import sharpy.utils.utils_ams as uams
 import sharpy.structure.utils.lagrangeconstraints as lagrangeconstraints
-# import matplotlib.pyplot as plt
 
 
 _BaseStructural = solver_from_string('_BaseStructural')
