### 2019-11-25 a15d11a
Travis now runs w/ the minimal environment
