### 2019-11-25 3755060
Added coverage installation after minimal env loaded
