### 2019-11-26 b04da14
Closes #40. Added a few modifications to competitor software in JOSS paper
