### 2019-11-27 5781ba2
Updated installation instructions
