+++ docs/source/content/installation.md
@@ -1,5 +1,5 @@
 # SHARPy Installation Guide
-__Last revision 15 November 2019__
+__Last revision 27 November 2019__
 
 The following step by step tutorial will guide you through the installation process of SHARPy.
 
@@ -87,10 +87,12 @@ python -m unittest
 
 **Enjoy!**
 
+## Building SHARPy from source (release or development builds)
 
-## Building SHARPy from source
+SHARPy can be built from source so that you can get the latest release or (stable) development build.
 
-SHARPy depends on the UVLM and xbeam repositories that are also found on GitHub:
+Apart from the SHARPy repository, you will need the UVLM and xbeam repositories that are also found on GitHub and are
+required to run SHARPy:
 
 + [xbeam](http://github.com/imperialcollegelondon/xbeam)
 + [UVLM](http://github.com/imperialcollegelondon/UVLM)
@@ -106,30 +108,50 @@ SHARPy depends on the UVLM and xbeam repositories that are also found on GitHub:
     ```bash
     git clone http://github.com/ImperialCollegeLondon/sharpy
     ```
-    To keep up with the latest development work, check out the `develop` branch:
-    ```bash
-    git checkout develop
-    ```
-    To run the release version of SHARPy, skip this last step.
+    
+    #### Building the release version of SHARPy. 
+    
+    1. Determine the latest version.
+    
+        The latest version can be found on the release bagde on the readme or in the documentation index page.
+        Alternatively, you can see SHARPy's release history from [Github releases](https://github.com/ImperialCollegeLondon/sharpy/releases).
+        
+    2. In your cloned `sharpy` directory, run:
+        
+        ```bash
+        git fetch && git fetch --tags
+        ```
+        This will pull the latest release tags, that are not normally pulled by default.
+        
+    3. Finally, check out the desired release tag. For instance,
+        ```bash
+        git checkout v1.0.1
+        ```
+    You are now running SHARPy's latest __release__ build. Continue with the installation of xbeam in step 3.
+    
+    #### Building the development version of SHARPy
+    
+    1. To keep up with the latest developments (stable build), check out the `master` branch:
+    
+        ```bash
+        git checkout master
+        ```
+       
+    2. Continue with the installation of `xbeam`.
+    
 2. Clone `xbeam` inside the working folder
     ```bash
     git clone http://github.com/ImperialCollegeLondon/xbeam
-    ```
-    Similarly, check out the `develop` branch for the latest work under development
-    ```bash
-    git checkout develop
     ```
 3. Clone the `UVLM` repository to the working directory
     ```bash
     git clone http://github.com/ImperialCollegeLondon/UVLM
     ```
-    Check out the `develop` branch for the latest work under development
-    ```bash
-    git checkout develop origin/develop
-    ```
-    Likewise, skip the `git checkout` commands if you are running the `master` branch in `sharpy`.
 
-### Set up the Python Environment
+SHARPy is now ready for you to compile its dependencies, just a few more steps...
+
+
+### Setting up the Python Environment
 
 SHARPy uses the Anaconda package manager to provide the necessary Python packages.
 These are specified in an Anaconda environment that shall be activated prior to compiling the xbeam and UVLM libraries
@@ -144,33 +166,22 @@ or running any SHARPy cases.
     #>>> Python 3.5.3 :: Anaconda custom (64-bit)
     ```
 
-        If it returns `Python 2.X.X` (where `X` does not matter), there are two possibilities:
[…]
