### 2019-11-27 b3d3584
A few more changes to install instructions
