+++ docs/source/includes/rom/frequencyresponseplot/index.rst
@@ -1,4 +1,0 @@
-.. toctree::
-	:glob:
-
-	./FrequencyResponseComparison
