+++ docs/source/includes/utils/exceptions/NotValidSetting.rst
@@ -1,0 +1,5 @@
+NotValidSetting
+---------------
+
+.. autoclass:: sharpy.utils.exceptions.NotValidSetting
+	:members:
