### 2019-11-27 ImperialCollegeLondon/sharpy#37 @fonsocarre
With #38 and #41, plus commit af5964a2f8a7a1e133c0948cb2fdeb5cb5c46568 this is now fixed.
