### 2019-11-27 904cbd0
Bug fixes in installation docs
