### 2019-11-27 ImperialCollegeLondon/sharpy#46 @rafmudaf
Installation fails on macOS with Intel compiler
I've successfully compiled and run SHARPy on Linux, but want to open an issue here for further discussion on compiling and running on macOS in regards to https://github.com/openjournals/joss-reviews/issues/1885.

At the moment, this issue is primarily dependent on https://github.com/ImperialCollegeLondon/UVLM/issues/4.
