### 2019-11-28 af84c6c
Fixes bug in Krylov dimension when using a single sided projection
