+++ sharpy/linear/src/surface.py
@@ -7,7 +7,6 @@ import ctypes as ct
 import numpy as np
 import itertools
 
-# from IPython import embed
 dmver = np.array([0, 1, 1, 0])  # delta to go from (m,n) panel to (m,n) vertices
 dnver = np.array([0, 0, 1, 1])
 
@@ -15,7 +14,7 @@ from sharpy.utils.sharpydir import SharpyDir
 import sharpy.utils.ctypes_utils as ct_utils
 import sharpy.linear.src.libuvlm as libuvlm
 
-libc = ct_utils.import_ctypes_lib(SharpyDir + '/lib/', 'libuvlm')
+libc = ct_utils.import_ctypes_lib(SharpyDir + '/lib/UVLM/lib/', 'libuvlm')
 
 
 class AeroGridGeo():
