+++ sharpy/structure/utils/xbeamlib.py
@@ -9,6 +9,7 @@ from sharpy.utils.sharpydir import SharpyDir
 # from sharpy.utils.datastructures import StructTimeStepInfo
 import sharpy.utils.cout_utils as cout
 
+xbeam_lib_path = SharpyDir + '/lib/xbeam/lib/'
 
 class Xbopts(ct.Structure):
     """Structure skeleton for options input in xbeam
@@ -60,7 +61,7 @@ class Xbopts(ct.Structure):
         self.relaxation_factor = ct.c_double(0.3)
 
 
-xbeamlib = ct_utils.import_ctypes_lib(SharpyDir + '/lib/', 'libxbeam')
+xbeamlib = ct_utils.import_ctypes_lib(xbeam_lib_path, 'libxbeam')
 
 # ctypes pointer types
 doubleP = ct.POINTER(ct.c_double)
@@ -795,7 +796,7 @@ def cbeam3_asbly_dynamic(beam, tstep, settings):
     """
 
     # library load
-    xbeamlib = ct_utils.import_ctypes_lib(SharpyDir + '/lib/', 'libxbeam')
+    xbeamlib = ct_utils.import_ctypes_lib(xbeam_lib_path, 'libxbeam')
     f_cbeam3_asbly_dynamic_python = xbeamlib.cbeam3_asbly_dynamic_python
     f_cbeam3_asbly_dynamic_python.restype = None
 
@@ -899,7 +900,7 @@ def xbeam3_asbly_dynamic(beam, tstep, settings):
     """
 
     # library load
-    xbeamlib = ct_utils.import_ctypes_lib(SharpyDir + '/lib/', 'libxbeam')
+    xbeamlib = ct_utils.import_ctypes_lib(xbeam_lib_path, 'libxbeam')
     f_xbeam3_asbly_dynamic_python = xbeamlib.xbeam3_asbly_dynamic_python
     f_xbeam3_asbly_dynamic_python.restype = None
 
@@ -992,7 +993,7 @@ def cbeam3_correct_gravity_forces(beam, tstep, settings):
     """
 
     # library load
-    xbeamlib = ct_utils.import_ctypes_lib(SharpyDir + '/lib/', 'libxbeam')
+    xbeamlib = ct_utils.import_ctypes_lib(xbeam_lib_path, 'libxbeam')
     f_cbeam3_correct_gravity_forces_python = xbeamlib.cbeam3_correct_gravity_forces_python
     f_cbeam3_correct_gravity_forces_python.restype = None
 
@@ -1046,7 +1047,7 @@ def cbeam3_asbly_static(beam, tstep, settings, iLoadStep):
     """
 
     # library load
-    xbeamlib = ct_utils.import_ctypes_lib(SharpyDir + '/lib/', 'libxbeam')
+    xbeamlib = ct_utils.import_ctypes_lib(xbeam_lib_path, 'libxbeam')
     f_cbeam3_asbly_static_python = xbeamlib.cbeam3_asbly_static_python
     f_cbeam3_asbly_static_python.restype = None
 
