### 2019-11-29 89dbd35
Update travis for submodules
