### 2019-12-01 189055f
Updated Dockr with submodule instructions
