### 2019-12-02 5cadd9b
Basic SHARPy installation script
