+++ install.sh
@@ -1,0 +1,9 @@
+!#/bin/sh
+source bin/sharpy_vars.sh
+git submodule init
+git submodule update
+cd lib/xbeam
+sh run_make.sh
+cd ../UVLM
+sh run_make.sh
+cd ../../
