### 2019-12-02 6633d75
Install instructions include new git submodules + Git branching model
