+++ docs/source/content/installation.md
@@ -1,5 +1,5 @@
 # SHARPy Installation Guide
-__Last revision 27 November 2019__
+__Last revision 2 December 2019__
 
 The following step by step tutorial will guide you through the installation process of SHARPy.
 
@@ -91,11 +91,9 @@ python -m unittest
 
 SHARPy can be built from source so that you can get the latest release or (stable) development build.
 
-Apart from the SHARPy repository, you will need the UVLM and xbeam repositories that are also found on GitHub and are
-required to run SHARPy:
-
-+ [xbeam](http://github.com/imperialcollegelondon/xbeam)
-+ [UVLM](http://github.com/imperialcollegelondon/UVLM)
+SHARPy depends on two external libraries, [xbeam](http://github.com/imperialcollegelondon/xbeam) and 
+[UVLM](http://github.com/imperialcollegelondon/UVLM). These are included as submodules to SHARPy and therefore
+once you initialise SHARPy you will also automatically clone the relevant versions of each library.
 
 ### Set up the folder structure
 
@@ -110,51 +108,8 @@ We will refer to this as the working folder.
     git clone http://github.com/ImperialCollegeLondon/sharpy
     ```
     
-#### Cloning the release version of SHARPy. 
-![Version badge](https://img.shields.io/endpoint.svg?url=https%3A%2F%2Fraw.githubusercontent.com%2FImperialCollegeLondon%2Fsharpy%2Fmaster%2F.version.json)
-
-1. Determine the latest version.
-
-    The latest version can be found on the release bagde on the readme or in the documentation index page.
-    Alternatively, you can see SHARPy's release history from [Github releases](https://github.com/ImperialCollegeLondon/sharpy/releases).
+2. We will now set up the SHARPy environment that will install other required distributions.
     
-2. In your cloned `sharpy` directory, run:
-    
-    ```bash
-    git fetch && git fetch --tags
-    ```
-    This will pull the latest release tags, that are not normally pulled by default.
-    
-3. Finally, check out the desired release tag. For instance,
-    ```bash
-    git checkout v1.0.1
-    ```
-You are now running SHARPy's latest __release__ build. 
-
-#### Cloning the development version of SHARPy
-
-1. To keep up with the latest developments (stable build), check out the `master` branch:
-
-    ```bash
-    git checkout master
-    ```
-   
-You are now running SHARPy's latest __development__ build. 
-
-#### Cloning xbeam and UVLM
-
-2. Clone `xbeam` inside the working folder
-    ```bash
-    git clone http://github.com/ImperialCollegeLondon/xbeam
-    ```
-3. Clone the `UVLM` repository to the working directory
-    ```bash
-    git clone http://github.com/ImperialCollegeLondon/UVLM
-    ```
-
-SHARPy is now ready for you to compile its dependencies, just a few more steps...
-
-
 ### Setting up the Python Environment
 
 SHARPy uses the Anaconda package manager to provide the necessary Python packages.
@@ -184,61 +139,96 @@ file if you are installing SHARPy on Mac OS X
     ```bash
     conda activate sharpy_env
     ```
-    you need to do this before you compile the `xbeam` and `uvlm` libs, as
-    some dependencies are included in the conda env.
+    you need to do this before you compile the `xbeam` and `uvlm` libraries, as
+    some dependencies are included in the conda environment.
+    
+    
+### Quick install
+The quick install is geared towards getting the release build of SHARPy running as quickly and simply as possible. If
+you would like to install a develop build or modify the compilation settings of the libraries skip to the next section.
 
-### Compiling the UVLM and xbeam libraries
-
-Once the folder structure has been laid out, the aerodynamic and structural libraries can be compiled.
[…]
