### 2019-12-03 a275eaa
Fixed typos in install instructions + changed geradin test to .sharpy
