+++ tests/xbeam/test_xbeam.py
@@ -21,11 +21,10 @@ class TestGeradinXbeam(unittest.TestCase):
         case = 'geradin'
         mod = importlib.import_module('tests.xbeam.' + case + '.generate_' + case)
 
-
     def test_geradin(self):
         import sharpy.sharpy_main
         # suppress screen output
-        solver_path = os.path.abspath(os.path.dirname(os.path.realpath(__file__)) + '/geradin/geradin.solver.txt')
+        solver_path = os.path.abspath(os.path.dirname(os.path.realpath(__file__)) + '/geradin/geradin.sharpy')
         sharpy.sharpy_main.main(['', solver_path])
 
         # read output and compare
@@ -41,7 +40,7 @@ class TestGeradinXbeam(unittest.TestCase):
     def tearDowns(self):
         solver_path = os.path.abspath(os.path.dirname(os.path.realpath(__file__)) + '/geradin/')
         files_to_delete = list()
-        extensions = ('*.txt', '*.h5')
+        extensions = ('*.txt', '*.h5', '*.sharpy')
         for f in extensions:
             files_to_delete.extend(glob.glob(solver_path + '/' + f))
 
