### 2019-12-03 2fa1a6a
Fix typo in unittests using tearDowns instead of tearDown
