### 2019-12-03 019f767
Added cd to move into sharpy before initialising submodules
