### 2019-12-03 f374c9b
CMakeLists added despite the gitignore
