### 2019-12-03 c4125ed
Git submodule commands automatic on cmake call
