### 2019-12-03 489c17f
travis updated for new cmake
