### 2019-12-03 73e428f
Updated env for MacOS, still needs to have more things added
