+++ utils/environment_macos.yml
@@ -1,118 +1,56 @@
 name: sharpy_env
 channels:
-  - conda-forge
   - defaults
 dependencies:
-  - alabaster=0.7.12=py37_0
-  - asn1crypto=1.2.0=py37_0
-  - attrs=19.3.0=py_0
-  - babel=2.7.0=py_0
   - blas=1.0=mkl
-  - bleach=3.1.0=py_0
-  - ca-certificates=2019.9.11=hecc5488_0
-  - certifi=2019.9.11=py37_0
-  - cffi=1.13.1=py37hb5b8e2f_0
-  - chardet=3.0.4=py37_1003
+  - ca-certificates=2019.11.27=0
+  - certifi=2019.11.28=py37_0
   - colorama=0.4.1=py37_0
-  - commonmark=0.9.0=py_0
-  - cryptography=2.8=py37ha12b0ac_0
-  - cycler=0.10.0=py37_0
-  - decorator=4.4.1=py_0
-  - defusedxml=0.6.0=py_0
+  - configobj=5.0.6=py37_1
+  - coverage=4.5.4=py37h1de35cc_0
   - dill=0.3.1.1=py37_0
-  - docutils=0.15.2=py37_0
   - eigen=3.3.7=h04f5b5a_0
-  - entrypoints=0.3=py37_1000
-  - freetype=2.9.1=hb4e5f40_0
-  - future=0.18.2=py37_0
-  - h5py=2.9.0=py37h3134771_0
-  - hdf5=1.10.4=hfa1e0ec_0
   - icu=58.2=h4b95b61_1
-  - idna=2.8=py37_0
-  - imagesize=1.1.0=py37_0
-  - importlib_metadata=0.23=py37_0
   - intel-openmp=2019.4=233
-  - ipython_genutils=0.2.0=py_1
-  - jinja2=2.10.3=py_0
-  - jsonschema=3.1.1=py37_0
-  - jupyter_core=4.5.0=py_0
-  - kiwisolver=1.1.0=py37h0a44026_0
-  - lapack=3.6.1=ha44fe06_2
   - libcxx=4.0.1=hcfea43d_1
   - libcxxabi=4.0.1=hcfea43d_1
   - libedit=3.1.20181209=hb402a30_0
   - libffi=3.2.1=h475c297_4
   - libgfortran=3.0.1=h93005f0_2
   - libiconv=1.15=hdd342a3_7
-  - libpng=1.6.37=ha441bb4_0
   - libxml2=2.9.9=hf6e021a_1
   - libxslt=1.1.33=h33a18ac_0
-  - lxml=4.4.1=py37hef8c89e_0
-  - markupsafe=1.1.1=py37h1de35cc_0
-  - matplotlib=3.1.1=py37h54f8f79_0
-  - mistune=0.8.4=py37h0b31af3_1000
+  - lxml=4.4.2=py37hef8c89e_0
   - mkl=2019.4=233
   - mkl-include=2019.4=233
   - mkl-service=2.3.0=py37hfbe908c_0
   - mkl_fft=1.0.15=py37h5e564d8_0
   - mkl_random=1.1.0=py37ha771720_0
-  - more-itertools=7.2.0=py_0
-  - nbconvert=5.6.1=py37_0
-  - nbformat=4.4.0=py_1
-  - nbsphinx=0.4.3=py_0
   - ncurses=6.1=h0a44026_1
-  - numpy=1.17.3=py37h4174a10_0
-  - numpy-base=1.17.3=py37h6575580_0
-  - openssl=1.1.1d=h0b31af3_0
-  - packaging=19.2=py_0
-  - pandas=0.25.2=py37h0a44026_0
-  - pandoc=2.7.3=0
[…]
