### 2019-12-04 a53f2ce
Updated MacOS env w/ matplotlib and hdf5
