### 2019-12-04 c553e69
Removed submodule init and update from CMakeLists
