+++ CMakeLists.txt
@@ -5,10 +5,10 @@ if(NOT CMAKE_BUILD_TYPE)
   set(CMAKE_BUILD_TYPE Release)
 endif()
 
-execute_process(
-    COMMAND git submodule init
-    WORKING_DIRECTORY ${CMAKE_SOURCE_DIR})
-execute_process(
-    COMMAND git submodule update
-    WORKING_DIRECTORY ${CMAKE_SOURCE_DIR})
+#execute_process(
+    #COMMAND git submodule init
+    #WORKING_DIRECTORY ${CMAKE_SOURCE_DIR})
+#execute_process(
+    #COMMAND git submodule update
+    #WORKING_DIRECTORY ${CMAKE_SOURCE_DIR})
 add_subdirectory(lib)
