### 2019-12-04 005210f
Updated Dockerfile for CMake
