### 2019-12-04 a28e8df
Added cmake to dockerfile
