### 2019-12-05 9778bcd
Added CMake to conda env instead of requiring it in the host
