+++ utils/environment_macos.yml
@@ -4,20 +4,25 @@ channels:
   - defaults
 dependencies:
   - blas=1.0=mkl
-  - ca-certificates=2019.11.27=0
+  - bzip2=1.0.8=h01d97ff_1
+  - ca-certificates=2019.11.28=hecc5488_0
   - certifi=2019.11.28=py37_0
+  - cmake=3.14.0=haff7e42_0
   - colorama=0.4.1=py37_0
   - configobj=5.0.6=py37_1
   - coverage=4.5.4=py37h1de35cc_0
   - cycler=0.10.0=py37_0
   - dill=0.3.1.1=py37_0
   - eigen=3.3.7=h04f5b5a_1000
+  - expat=2.2.6=h0a44026_0
   - freetype=2.9.1=hb4e5f40_0
   - h5py=2.9.0=py37h3134771_0
   - hdf5=1.10.4=hfa1e0ec_0
   - icu=58.2=h4b95b61_1
   - intel-openmp=2019.4=233
   - kiwisolver=1.1.0=py37h0a44026_0
+  - krb5=1.16.1=hddcf347_7
+  - libcurl=7.67.0=h051b688_0
   - libcxx=4.0.1=hcfea43d_1
   - libcxxabi=4.0.1=hcfea43d_1
   - libedit=3.1.20181209=hb402a30_0
@@ -25,6 +30,7 @@ dependencies:
   - libgfortran=3.0.1=h93005f0_2
   - libiconv=1.15=hdd342a3_7
   - libpng=1.6.37=ha441bb4_0
+  - libssh2=1.8.2=hcdc9a53_2
   - libxml2=2.9.9=hf6e021a_1
   - libxslt=1.1.33=h33a18ac_0
   - lxml=4.4.2=py37hef8c89e_0
@@ -37,7 +43,7 @@ dependencies:
   - ncurses=6.1=h0a44026_1
   - numpy=1.17.4=py37h890c691_0
   - numpy-base=1.17.4=py37h6575580_0
-  - openssl=1.1.1d=h1de35cc_3
+  - openssl=1.1.1d=h0b31af3_0
   - pandas=0.25.3=py37h0a44026_0
   - pip=19.3.1=py37_0
   - pyparsing=2.4.5=py_0
@@ -45,6 +51,7 @@ dependencies:
   - python-dateutil=2.8.1=py_0
   - pytz=2019.3=py_0
   - readline=7.0=h1de35cc_5
+  - rhash=1.3.8=ha12b0ac_0
   - scipy=1.3.1=py37h1410ff5_0
   - setuptools=42.0.2=py37_0
   - six=1.13.0=py37_0
@@ -52,7 +59,7 @@ dependencies:
   - tk=8.6.8=ha441bb4_0
   - tornado=6.0.3=py37h1de35cc_0
   - wheel=0.33.6=py37_0
-  - xlrd=1.2.0
+  - xlrd=1.2.0=py_0
   - xz=5.2.4=h1de35cc_4
   - zlib=1.2.11=h1de35cc_3
   - pip:
