### 2019-12-05 811c442
Added Cmake to linux conda envs
