+++ utils/environment_minimal.yml
@@ -5,22 +5,28 @@ channels:
 dependencies:
   - _libgcc_mutex=0.1=main
   - blas=1.0=mkl
-  - ca-certificates=2019.10.16=0
-  - certifi=2019.9.11=py37_0
+  - bzip2=1.0.8=h7b6447c_0
+  - ca-certificates=2019.11.27=0
+  - certifi=2019.11.28=py37_0
+  - cmake=3.14.0=h52cb24c_0
   - colorama=0.4.1=py37_0
   - configobj=5.0.6=py37_1
   - dill=0.3.1.1=py37_0
   - eigen=3.3.7=hfd86e86_0
+  - expat=2.2.6=he6710b0_0
   - h5py=2.9.0=py37h7918eee_0
   - hdf5=1.10.4=hb1b8bf9_0
   - icu=58.2=h9c2bf20_1
   - intel-openmp=2019.4=243
+  - krb5=1.16.1=h173b8e3_7
   - lapack=3.6.1=ha44fe06_2
+  - libcurl=7.67.0=h20c2e04_0
   - libedit=3.1.20181209=hc058e9b_0
   - libffi=3.2.1=hd88cf55_4
   - libgcc-ng=9.1.0=hdf63c60_0
   - libgfortran=3.0.0=1
   - libgfortran-ng=7.3.0=hdf63c60_0
+  - libssh2=1.8.2=h1ba5d50_0
   - libstdcxx-ng=9.1.0=hdf63c60_0
   - libxml2=2.9.9=hea5a465_1
   - libxslt=1.1.33=h7d1a2b0_0
@@ -41,6 +47,7 @@ dependencies:
   - pytz=2019.3=py_0
   - pyyaml=5.1.2=py37h7b6447c_0
   - readline=7.0=h7b6447c_5
+  - rhash=1.3.8=h1ba5d50_0
   - scipy=1.3.1=py37h7c811a0_0
   - setuptools=41.6.0=py37_0
   - six=1.13.0=py37_0
