+++ utils/environment_linux.yml
@@ -10,15 +10,18 @@ dependencies:
   - babel=2.7.0=py_0
   - blas=1.0=mkl
   - bleach=3.1.0=py_0
-  - ca-certificates=2019.9.11=hecc5488_0
-  - certifi=2019.9.11=py37_0
+  - bzip2=1.0.8=h516909a_1
+  - ca-certificates=2019.11.27=0
+  - certifi=2019.11.28=py37_0
   - cffi=1.13.1=py37h2e261b9_0
   - chardet=3.0.4=py37_1003
+  - cmake=3.14.0=h52cb24c_0
   - colorama=0.4.1=py37_0
   - commonmark=0.9.0=py_0
   - configobj=5.0.6=py37_1
-  - coverage
+  - coverage=4.5.4=py37h516909a_0
   - cryptography=2.8=py37h1ba5d50_0
+  - curl=7.67.0=hbc83047_0
   - cycler=0.10.0=py37_0
   - dbus=1.13.12=h746ee38_0
   - decorator=4.4.1=py_0
@@ -35,6 +38,7 @@ dependencies:
   - gst-plugins-base=1.14.0=hbbd80ab_1
   - gstreamer=1.14.0=hb453b48_1
   - h5py=2.9.0=py37h7918eee_0
+  - hdf4=4.2.13=0
   - hdf5=1.10.4=hb1b8bf9_0
   - icu=58.2=h9c2bf20_1
   - idna=2.8=py37_0
@@ -44,22 +48,33 @@ dependencies:
   - ipython_genutils=0.2.0=py_1
   - jinja2=2.10.3=py_0
   - jpeg=9b=h024ee3a_2
+  - jsoncpp=1.8.4=hc9558a2_1002
   - jsonschema=3.1.1=py37_0
   - jupyter_core=4.5.0=py_0
   - kiwisolver=1.1.0=py37he6710b0_0
+  - krb5=1.16.1=h173b8e3_7
   - lapack=3.6.1=ha44fe06_2
+  - libcurl=7.67.0=h20c2e04_0
   - libedit=3.1.20181209=hc058e9b_0
   - libffi=3.2.1=hd88cf55_4
   - libgcc-ng=9.1.0=hdf63c60_0
   - libgfortran=3.0.0=1
   - libgfortran-ng=7.3.0=hdf63c60_0
+  - libnetcdf=4.6.1=h11d0813_2
+  - libogg=1.3.2=h14c3975_1001
   - libpng=1.6.37=hbc83047_0
+  - libssh2=1.8.2=h22169c7_2
   - libstdcxx-ng=9.1.0=hdf63c60_0
+  - libtheora=1.1.1=h516909a_1002
+  - libtiff=4.1.0=h2733197_0
   - libuuid=1.0.3=h1bed415_2
+  - libvorbis=1.3.5=h14c3975_1001
   - libxcb=1.13=h1bed415_1
   - libxml2=2.9.9=hea5a465_1
   - libxslt=1.1.33=h7d1a2b0_0
   - lxml=4.4.1=py37hefd8a0e_0
+  - lz4=2.2.1=py37hd79334b_0
+  - lz4-c=1.8.3=he1b5a44_1001
   - markupsafe=1.1.1=py37h7b6447c_0
   - matplotlib=3.1.1=py37h5429711_0
   - mistune=0.8.4=py37h516909a_1000
@@ -75,7 +90,7 @@ dependencies:
   - ncurses=6.1=he6710b0_1
   - numpy=1.17.3=py37hd14ec0e_0
   - numpy-base=1.17.3=py37hde5b4d6_0
-  - openssl=1.1.1d=h516909a_0
+  - openssl=1.1.1d=h7b6447c_3
[…]
