### 2019-12-05 f31ccc6
Removed cmake from Dockerfile
