+++ Dockerfile
@@ -7,7 +7,7 @@ ENV PATH=${PATH}:/miniconda3/bin
 
 # Development tools including compilers
 RUN yum groupinstall "Development Tools" -y --nogpgcheck && \
-    yum install -y --nogpgcheck mesa-libGL libXt libXt-devel wget gcc-gfortran lapack vim tmux cmake && \
+    yum install -y --nogpgcheck mesa-libGL libXt libXt-devel wget gcc-gfortran lapack vim tmux && \
     yum clean all
 
 # Install miniconda
