### 2019-12-05 fc0a084
Dockerfile updated and working
