+++ Dockerfile
@@ -16,8 +16,7 @@ RUN wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -
     /miniconda.sh -b -p /miniconda3/ && \
     rm /miniconda.sh && hash -r
 
-# Get SHARPy
-RUN git clone --recursive https://github.com/imperialcollegelondon/sharpy
+ADD / /sharpy_dir/
 
 # Update conda and make it run with no user interaction
 # Cleanup conda installation
@@ -25,15 +24,16 @@ RUN conda init bash && \
     conda config --set always_yes yes --set changeps1 no && \
     conda update -q conda && \
     conda config --set auto_activate_base false && \
-    conda env create -f sharpy/utils/environment_minimal.yml && conda clean -afy && \
+    conda env create -f /sharpy_dir/utils/environment_minimal.yml && conda clean -afy && \
     find /miniconda3/ -follow -type f -name '*.a' -delete && \
     find /miniconda3/ -follow -type f -name '*.pyc' -delete && \
     find /miniconda3/ -follow -type f -name '*.js.map' -delete
 
-COPY ./utils/docker/* /root/
+#COPY /utils/docker/* /root/
+RUN ln -s /sharpy_dir/utils/docker/* /root/
 
-RUN cd /sharpy && conda activate sharpy_minimal && \
-    mkdir build && cd build && cmake .. && make install -j 2 \
+RUN cd sharpy_dir && conda activate sharpy_minimal && \
+    mkdir build && cd build && CXX=g++ FC=gfortran cmake .. && make install -j 2 \
     && cd .. && rm -rf build
 
 ENTRYPOINT ["/bin/bash", "--init-file", "/root/bashrc"]
