### 2019-12-05 ImperialCollegeLondon/sharpy#50 @petebachant
WindTurbine case generation script does not produce sharpy file
Running in Docker container for https://github.com/openjournals/joss-reviews/issues/1885.

The HDF5 files appear to be created correctly, but I don't get a case file.
