### 2019-12-06 #1885 @petebachant
I noticed a small error in one of the sample cases, but was able to confirm the functionality of the software with a different one. This submission is acceptable to me.
