### 2019-12-06 ImperialCollegeLondon/sharpy#50 @ArturoMS13
Hi Pete,

If the problem is with `coupled/WindTubine` I get your point because it generates the file with the legacy name `rotor.solver.txt`. I will update it to `rotor.sharpy`. The legacy case still works as `sharpy rotor.solver.txt`.

If it is with another case, please let me know.

Thank you!
