### 2019-12-06 7a29dac
Added submodule init & update in Dockerfile
