+++ CMakeLists.txt
@@ -1,4 +1,4 @@
-cake_minimum_required(VERSION 3.9)
+cmake_minimum_required(VERSION 3.9)
 project(sharpy)
 
 if(NOT CMAKE_BUILD_TYPE)
