### 2019-12-06 97be2c7
change .solver.txt to .sharpy extension
