+++ tests/xbeam/geradin/generate_geradin.py
@@ -12,7 +12,7 @@ def clean_test_files():
     if os.path.isfile(fem_file_name):
         os.remove(fem_file_name)
 
-    solver_file_name = route + '/' + case_name + '.solver.txt'
+    solver_file_name = route + '/' + case_name + '.sharpy'
     if os.path.isfile(solver_file_name):
         os.remove(solver_file_name)
 
@@ -132,7 +132,7 @@ def generate_fem_file(route, case_name, num_elem, num_node_elem=3):
 
 
 def generate_solver_file():
-    file_name = route + '/' + case_name + '.solver.txt'
+    file_name = route + '/' + case_name + '.sharpy'
     # config = configparser.ConfigParser()
     import configobj
     config = configobj.ConfigObj()
