+++ cases/coupled/hinged_controlled_wing/generate_hinged_controlled_wing.py
@@ -73,7 +73,7 @@ def clean_test_files():
     if os.path.isfile(dyn_file_name):
         os.remove(dyn_file_name)
 
-    solver_file_name = route + '/' + case_name + '.solver.txt'
+    solver_file_name = route + '/' + case_name + '.sharpy'
     if os.path.isfile(solver_file_name):
         os.remove(solver_file_name)
 
@@ -394,7 +394,7 @@ def generate_multibody_file():
     gc.generate_multibody_file(LC, MB, route, case_name)
 
 def generate_solver_file(horseshoe=False):
-    file_name = route + '/' + case_name + '.solver.txt'
+    file_name = route + '/' + case_name + '.sharpy'
     # config = configparser.ConfigParser()
     import configobj
     config = configobj.ConfigObj()
