### 2019-12-06 4699745
Environment_linux does not include MKL for compatibility
