+++ utils/environment_linux.yml
@@ -8,20 +8,16 @@ dependencies:
   - asn1crypto=1.2.0=py37_0
   - attrs=19.3.0=py_0
   - babel=2.7.0=py_0
-  - blas=1.0=mkl
+  - backcall=0.1.0=py37_0
   - bleach=3.1.0=py_0
-  - bzip2=1.0.8=h516909a_1
   - ca-certificates=2019.11.27=0
   - certifi=2019.11.28=py37_0
   - cffi=1.13.1=py37h2e261b9_0
   - chardet=3.0.4=py37_1003
-  - cmake=3.14.0=h52cb24c_0
   - colorama=0.4.1=py37_0
   - commonmark=0.9.0=py_0
   - configobj=5.0.6=py37_1
-  - coverage=4.5.4=py37h516909a_0
   - cryptography=2.8=py37h1ba5d50_0
-  - curl=7.67.0=hbc83047_0
   - cycler=0.10.0=py37_0
   - dbus=1.13.12=h746ee38_0
   - decorator=4.4.1=py_0
@@ -37,66 +33,67 @@ dependencies:
   - glib=2.63.1=h5a9c865_0
   - gst-plugins-base=1.14.0=hbbd80ab_1
   - gstreamer=1.14.0=hb453b48_1
-  - h5py=2.9.0=py37h7918eee_0
-  - hdf4=4.2.13=0
+  - h5py=2.9.0=nompi_py37hcafd542_1103
   - hdf5=1.10.4=hb1b8bf9_0
   - icu=58.2=h9c2bf20_1
   - idna=2.8=py37_0
   - imagesize=1.1.0=py37_0
   - importlib_metadata=0.23=py37_0
   - intel-openmp=2019.4=243
+  - ipykernel=5.1.3=py37h39e3cac_0
+  - ipython=7.10.1=py37h39e3cac_0
   - ipython_genutils=0.2.0=py_1
+  - jedi=0.15.1=py37_0
   - jinja2=2.10.3=py_0
   - jpeg=9b=h024ee3a_2
-  - jsoncpp=1.8.4=hc9558a2_1002
+  - json5=0.8.5=py_0
   - jsonschema=3.1.1=py37_0
+  - jupyter_client=5.3.3=py37_1
   - jupyter_core=4.5.0=py_0
+  - jupyterlab=1.2.3=py_0
+  - jupyterlab_server=1.0.6=py_0
   - kiwisolver=1.1.0=py37he6710b0_0
-  - krb5=1.16.1=h173b8e3_7
   - lapack=3.6.1=ha44fe06_2
-  - libcurl=7.67.0=h20c2e04_0
+  - libblas=3.8.0=14_openblas
+  - libcblas=3.8.0=14_openblas
   - libedit=3.1.20181209=hc058e9b_0
   - libffi=3.2.1=hd88cf55_4
   - libgcc-ng=9.1.0=hdf63c60_0
   - libgfortran=3.0.0=1
   - libgfortran-ng=7.3.0=hdf63c60_0
-  - libnetcdf=4.6.1=h11d0813_2
-  - libogg=1.3.2=h14c3975_1001
+  - liblapack=3.8.0=14_openblas
+  - libopenblas=0.3.7=h5ec1e0e_4
   - libpng=1.6.37=hbc83047_0
-  - libssh2=1.8.2=h22169c7_2
+  - libsodium=1.0.16=h1bed415_0
   - libstdcxx-ng=9.1.0=hdf63c60_0
-  - libtheora=1.1.1=h516909a_1002
-  - libtiff=4.1.0=h2733197_0
   - libuuid=1.0.3=h1bed415_2
[…]
