### 2019-12-06 e555383
Environment_minimal has no MKL included anymore
