+++ utils/environment_minimal.yml
@@ -4,67 +4,81 @@ channels:
   - defaults
 dependencies:
   - _libgcc_mutex=0.1=main
-  - blas=1.0=mkl
-  - bzip2=1.0.8=h7b6447c_0
   - ca-certificates=2019.11.27=0
   - certifi=2019.11.28=py37_0
-  - cmake=3.14.0=h52cb24c_0
   - colorama=0.4.1=py37_0
-  - configobj=5.0.6=py37_1
+  - cycler=0.10.0=py_2
+  - dbus=1.13.6=he372182_0
   - dill=0.3.1.1=py37_0
-  - eigen=3.3.7=hfd86e86_0
-  - expat=2.2.6=he6710b0_0
+  - eigen=3.3.7=hc9558a2_1001
+  - expat=2.2.5=he1b5a44_1004
+  - fontconfig=2.13.1=he4413a7_1000
+  - freetype=2.10.0=he983fc9_1
+  - gettext=0.19.8.1=hc5be6a0_1002
+  - glib=2.58.3=py37h6f030ca_1002
+  - gst-plugins-base=1.14.5=h0935bb2_0
+  - gstreamer=1.14.5=h36ae1b5_0
   - h5py=2.9.0=py37h7918eee_0
   - hdf5=1.10.4=hb1b8bf9_0
-  - icu=58.2=h9c2bf20_1
-  - intel-openmp=2019.4=243
-  - krb5=1.16.1=h173b8e3_7
+  - icu=58.2=hf484d3e_1000
+  - jpeg=9c=h14c3975_1001
+  - kiwisolver=1.1.0=py37hc9558a2_0
   - lapack=3.6.1=ha44fe06_2
-  - libcurl=7.67.0=h20c2e04_0
+  - libblas=3.8.0=14_openblas
+  - libcblas=3.8.0=14_openblas
   - libedit=3.1.20181209=hc058e9b_0
   - libffi=3.2.1=hd88cf55_4
   - libgcc-ng=9.1.0=hdf63c60_0
   - libgfortran=3.0.0=1
-  - libgfortran-ng=7.3.0=hdf63c60_0
-  - libssh2=1.8.2=h1ba5d50_0
+  - libgfortran-ng=7.3.0=hdf63c60_2
+  - libiconv=1.15=h516909a_1005
+  - liblapack=3.8.0=14_openblas
+  - libopenblas=0.3.7=h5ec1e0e_4
+  - libpng=1.6.37=hed695b0_0
   - libstdcxx-ng=9.1.0=hdf63c60_0
+  - libuuid=2.32.1=h14c3975_1000
+  - libxcb=1.13=h14c3975_1002
   - libxml2=2.9.9=hea5a465_1
   - libxslt=1.1.33=h7d1a2b0_0
-  - lxml=4.4.1=py37hefd8a0e_0
-  - mkl=2019.4=243
-  - mkl-include=2019.4=243
-  - mkl-service=2.3.0=py37he904b0f_0
-  - mkl_fft=1.0.15=py37ha843d7b_0
-  - mkl_random=1.1.0=py37hd6b4f25_0
+  - lxml=4.4.2=py37hefd8a0e_0
+  - matplotlib=3.1.1=py37h5429711_0
   - ncurses=6.1=he6710b0_1
-  - numpy=1.17.3=py37hd14ec0e_0
-  - numpy-base=1.17.3=py37hde5b4d6_0
+  - numpy=1.17.3=py37h95a1406_0
   - openssl=1.1.1d=h7b6447c_3
-  - pandas=0.25.2=py37he6710b0_0
+  - pandas=0.25.3=py37hb3f55d8_0
+  - pcre=8.43=he1b5a44_0
   - pip=19.3.1=py37_0
[…]
