### 2019-12-06 6923e7b
Env minimal does not rely on pip any more
