+++ utils/environment_minimal.yml
@@ -4,28 +4,38 @@ channels:
   - defaults
 dependencies:
   - _libgcc_mutex=0.1=main
+  - apptools=4.5.0=py_0
+  - bzip2=1.0.8=h7b6447c_0
   - ca-certificates=2019.11.27=0
   - certifi=2019.11.28=py37_0
   - colorama=0.4.1=py37_0
+  - configobj=5.0.6=py_0
+  - curl=7.67.0=hbc83047_0
   - cycler=0.10.0=py_2
   - dbus=1.13.6=he372182_0
   - dill=0.3.1.1=py37_0
   - eigen=3.3.7=hc9558a2_1001
-  - expat=2.2.5=he1b5a44_1004
+  - envisage=4.8.0=py_0
+  - expat=2.2.6=he6710b0_0
   - fontconfig=2.13.1=he4413a7_1000
   - freetype=2.10.0=he983fc9_1
+  - future=0.18.2=py37_0
   - gettext=0.19.8.1=hc5be6a0_1002
   - glib=2.58.3=py37h6f030ca_1002
   - gst-plugins-base=1.14.5=h0935bb2_0
   - gstreamer=1.14.5=h36ae1b5_0
   - h5py=2.9.0=py37h7918eee_0
+  - hdf4=4.2.13=h3ca952b_2
   - hdf5=1.10.4=hb1b8bf9_0
   - icu=58.2=hf484d3e_1000
   - jpeg=9c=h14c3975_1001
+  - jsoncpp=1.8.4=hfd86e86_0
   - kiwisolver=1.1.0=py37hc9558a2_0
+  - krb5=1.16.1=h173b8e3_7
   - lapack=3.6.1=ha44fe06_2
   - libblas=3.8.0=14_openblas
   - libcblas=3.8.0=14_openblas
+  - libcurl=7.67.0=h20c2e04_0
   - libedit=3.1.20181209=hc058e9b_0
   - libffi=3.2.1=hd88cf55_4
   - libgcc-ng=9.1.0=hdf63c60_0
@@ -33,15 +43,23 @@ dependencies:
   - libgfortran-ng=7.3.0=hdf63c60_2
   - libiconv=1.15=h516909a_1005
   - liblapack=3.8.0=14_openblas
+  - libnetcdf=4.6.1=h11d0813_2
+  - libogg=1.3.2=h7b6447c_0
   - libopenblas=0.3.7=h5ec1e0e_4
   - libpng=1.6.37=hed695b0_0
+  - libssh2=1.8.2=h1ba5d50_0
   - libstdcxx-ng=9.1.0=hdf63c60_0
+  - libtheora=1.1.1=h5ab3b9f_1
+  - libtiff=4.1.0=h2733197_0
   - libuuid=2.32.1=h14c3975_1000
+  - libvorbis=1.3.6=h7b6447c_0
   - libxcb=1.13=h14c3975_1002
   - libxml2=2.9.9=hea5a465_1
   - libxslt=1.1.33=h7d1a2b0_0
   - lxml=4.4.2=py37hefd8a0e_0
+  - lz4-c=1.8.1.2=h14c3975_0
   - matplotlib=3.1.1=py37h5429711_0
+  - mayavi=4.6.2=py37hcf37d21_4
   - ncurses=6.1=he6710b0_1
   - numpy=1.17.3=py37h95a1406_0
   - openssl=1.1.1d=h7b6447c_3
@@ -49,6 +67,8 @@ dependencies:
   - pcre=8.43=he1b5a44_0
   - pip=19.3.1=py37_0
   - pthread-stubs=0.4=h14c3975_1001
+  - pyface=6.1.2=py37_0
[…]
