### 2019-12-06 d389bc0
Removed pip dependency for managing packages
