+++ utils/environment_linux.yml
@@ -5,11 +5,13 @@ channels:
 dependencies:
   - _libgcc_mutex=0.1=main
   - alabaster=0.7.12=py37_0
+  - apptools=4.4.0=py37_1
   - asn1crypto=1.2.0=py37_0
   - attrs=19.3.0=py_0
   - babel=2.7.0=py_0
   - backcall=0.1.0=py37_0
   - bleach=3.1.0=py_0
+  - bzip2=1.0.8=h7b6447c_0
   - ca-certificates=2019.11.27=0
   - certifi=2019.11.28=py37_0
   - cffi=1.13.1=py37h2e261b9_0
@@ -18,6 +20,7 @@ dependencies:
   - commonmark=0.9.0=py_0
   - configobj=5.0.6=py37_1
   - cryptography=2.8=py37h1ba5d50_0
+  - curl=7.67.0=hbc83047_0
   - cycler=0.10.0=py37_0
   - dbus=1.13.12=h746ee38_0
   - decorator=4.4.1=py_0
@@ -26,6 +29,7 @@ dependencies:
   - docutils=0.15.2=py37_0
   - eigen=3.3.7=hfd86e86_0
   - entrypoints=0.3=py37_1000
+  - envisage=4.8.0=py_0
   - expat=2.2.6=he6710b0_0
   - fontconfig=2.13.0=h9420a91_0
   - freetype=2.9.1=h8a8886c_1
@@ -34,6 +38,7 @@ dependencies:
   - gst-plugins-base=1.14.0=hbbd80ab_1
   - gstreamer=1.14.0=hb453b48_1
   - h5py=2.9.0=nompi_py37hcafd542_1103
+  - hdf4=4.2.13=h3ca952b_2
   - hdf5=1.10.4=hb1b8bf9_0
   - icu=58.2=h9c2bf20_1
   - idna=2.8=py37_0
@@ -47,32 +52,43 @@ dependencies:
   - jinja2=2.10.3=py_0
   - jpeg=9b=h024ee3a_2
   - json5=0.8.5=py_0
+  - jsoncpp=1.8.4=hfd86e86_0
   - jsonschema=3.1.1=py37_0
   - jupyter_client=5.3.3=py37_1
   - jupyter_core=4.5.0=py_0
   - jupyterlab=1.2.3=py_0
   - jupyterlab_server=1.0.6=py_0
   - kiwisolver=1.1.0=py37he6710b0_0
+  - krb5=1.16.1=h173b8e3_7
   - lapack=3.6.1=ha44fe06_2
   - libblas=3.8.0=14_openblas
   - libcblas=3.8.0=14_openblas
+  - libcurl=7.67.0=h20c2e04_0
   - libedit=3.1.20181209=hc058e9b_0
   - libffi=3.2.1=hd88cf55_4
   - libgcc-ng=9.1.0=hdf63c60_0
   - libgfortran=3.0.0=1
   - libgfortran-ng=7.3.0=hdf63c60_0
   - liblapack=3.8.0=14_openblas
+  - libnetcdf=4.6.1=h11d0813_2
+  - libogg=1.3.2=h7b6447c_0
   - libopenblas=0.3.7=h5ec1e0e_4
   - libpng=1.6.37=hbc83047_0
   - libsodium=1.0.16=h1bed415_0
+  - libssh2=1.8.2=h1ba5d50_0
   - libstdcxx-ng=9.1.0=hdf63c60_0
+  - libtheora=1.1.1=h5ab3b9f_1
+  - libtiff=4.1.0=h2733197_0
   - libuuid=1.0.3=h1bed415_2
+  - libvorbis=1.3.6=h7b6447c_0
[…]
