### 2019-12-06 b700f14
Added CMake to minimal env
