+++ utils/environment_minimal.yml
@@ -8,6 +8,7 @@ dependencies:
   - bzip2=1.0.8=h7b6447c_0
   - ca-certificates=2019.11.27=0
   - certifi=2019.11.28=py37_0
+  - cmake=3.14.0=h52cb24c_0
   - colorama=0.4.1=py37_0
   - configobj=5.0.6=py_0
   - curl=7.67.0=hbc83047_0
@@ -77,6 +78,7 @@ dependencies:
   - pyyaml=5.1.2=py37h7b6447c_0
   - qt=5.9.7=h52cfd70_2
   - readline=7.0=h7b6447c_5
+  - rhash=1.3.8=h1ba5d50_0
   - scipy=1.3.2=py37h921218d_0
   - setuptools=42.0.2=py37_0
   - sip=4.19.8=py37hf484d3e_0
