### 2019-12-06 3503c85
Removed old install.sh
