### 2019-12-07 b0bc6ad
environment_linux now includes cmake
