+++ utils/environment_linux.yml
@@ -16,6 +16,7 @@ dependencies:
   - certifi=2019.11.28=py37_0
   - cffi=1.13.1=py37h2e261b9_0
   - chardet=3.0.4=py37_1003
+  - cmake=3.14.0=h52cb24c_0
   - colorama=0.4.1=py37_0
   - commonmark=0.9.0=py_0
   - configobj=5.0.6=py37_1
@@ -127,6 +128,7 @@ dependencies:
   - readline=7.0=h7b6447c_5
   - recommonmark=0.6.0=py_0
   - requests=2.22.0=py37_0
+  - rhash=1.3.8=h1ba5d50_0
   - scipy=1.3.2=py37h921218d_0
   - send2trash=1.5.0=py37_0
   - setuptools=41.6.0=py37_0
@@ -162,3 +164,4 @@ dependencies:
   - zipp=0.6.0=py_0
   - zlib=1.2.11=h7b6447c_3
   - zstd=1.3.7=h0b5b093_0
+
