### 2019-12-07 6502c79
Added cmake and remove pip dependencies in macOS env
