+++ utils/environment_macos.yml
@@ -3,23 +3,33 @@ channels:
   - conda-forge
   - defaults
 dependencies:
+  - apptools=4.4.0=py37_1
   - blas=1.0=mkl
   - bzip2=1.0.8=h01d97ff_1
-  - ca-certificates=2019.11.28=hecc5488_0
+  - ca-certificates=2019.11.27=0
   - certifi=2019.11.28=py37_0
   - cmake=3.14.0=haff7e42_0
   - colorama=0.4.1=py37_0
   - configobj=5.0.6=py37_1
   - coverage=4.5.4=py37h1de35cc_0
+  - curl=7.67.0=ha441bb4_0
   - cycler=0.10.0=py37_0
+  - dbus=1.13.12=h90a0687_0
   - dill=0.3.1.1=py37_0
   - eigen=3.3.7=h04f5b5a_1000
+  - envisage=4.8.0=py_0
   - expat=2.2.6=h0a44026_0
   - freetype=2.9.1=hb4e5f40_0
+  - future=0.18.2=py37_0
+  - gettext=0.19.8.1=h15daf44_3
+  - glib=2.63.1=hd977a24_0
   - h5py=2.9.0=py37h3134771_0
+  - hdf4=4.2.13=h39711bb_2
   - hdf5=1.10.4=hfa1e0ec_0
   - icu=58.2=h4b95b61_1
   - intel-openmp=2019.4=233
+  - jpeg=9b=he5867d9_2
+  - jsoncpp=1.8.4=h04f5b5a_0
   - kiwisolver=1.1.0=py37h0a44026_0
   - krb5=1.16.1=hddcf347_7
   - libcurl=7.67.0=h051b688_0
@@ -29,12 +39,19 @@ dependencies:
   - libffi=3.2.1=h475c297_4
   - libgfortran=3.0.1=h93005f0_2
   - libiconv=1.15=hdd342a3_7
+  - libnetcdf=4.6.1=hd5207e6_2
+  - libogg=1.3.2=h1de35cc_0
   - libpng=1.6.37=ha441bb4_0
   - libssh2=1.8.2=hcdc9a53_2
+  - libtheora=1.1.1=hb4e5f40_1
+  - libtiff=4.1.0=hcb84e12_0
+  - libvorbis=1.3.6=h1de35cc_0
   - libxml2=2.9.9=hf6e021a_1
   - libxslt=1.1.33=h33a18ac_0
   - lxml=4.4.2=py37hef8c89e_0
+  - lz4-c=1.8.1.2=h1de35cc_0
   - matplotlib=3.1.1=py37h54f8f79_0
+  - mayavi=4.6.2=py37hdde6e19_4
   - mkl=2019.4=233
   - mkl-include=2019.4=233
   - mkl-service=2.3.0=py37hfbe908c_0
@@ -43,32 +60,34 @@ dependencies:
   - ncurses=6.1=h0a44026_1
   - numpy=1.17.4=py37h890c691_0
   - numpy-base=1.17.4=py37h6575580_0
-  - openssl=1.1.1d=h0b31af3_0
+  - openssl=1.1.1d=h1de35cc_3
   - pandas=0.25.3=py37h0a44026_0
+  - pcre=8.43=h0a44026_0
   - pip=19.3.1=py37_0
+  - pyface=6.1.2=py37_0
+  - pygments=2.5.2=py_0
   - pyparsing=2.4.5=py_0
+  - pyqt=5.9.2=py37h655552a_2
   - python=3.7.5=h359304d_0
   - python-dateutil=2.8.1=py_0
   - pytz=2019.3=py_0
+  - qt=5.9.7=h468cd18_1
[…]
