### 2019-12-07 39c7deb
Updated installation instructions
Removed CMake from requirements since it is included in the environment
