### 2019-12-08 #1885 @rafmudaf
After the latest updates to the build system, I was able to compile on macOS exactly as documented. I also was able to run the example cases and confirm the functionality. Very good work!
