### 2019-12-08 #1885 @labarba
@fonsocarre — small typo: ¶4 "These longer and much slender…" >> These longer and more slender...

Also, please use consistent capitalization for your references: normal cap for article titles, camel case for journal or conference title. The first two references need proper capitalization of the conference (scitech forum >> SciTech Forum), and the 7th reference needs the article title in normal cap.

Once you make these small changes, please go ahead and make a new tagged release of your software, and report the version number here, then archive your software repository in Zenodo (or similar) and please report the DOI here.
