### 2019-12-08 #1885 @ngoiz
Thanks @labarba, @petebachant and @rafmudaf for the review and suggestions!

We will update the JOSS submission with the above, release and share the DOI.

Thanks!
