### 2019-12-09 356ca8f
Fixed typos in paper and bib
