+++ docs/JOSS/paper.bib
@@ -223,7 +223,7 @@ journal = {AIAA Journal},
 month = {mar},
 number = {3},
 pages = {528--538},
-title = {{Consistent Structural Linearization in Flexible Aircraft Dynamics with Large Rigid-Body Motion}},
+title = {Consistent Structural Linearization in Flexible Aircraft Dynamics with Large Rigid-Body Motion},
 volume = {52},
 year = {2014}
 }
@@ -348,7 +348,7 @@ year = {2012}
  @inproceedings{del2019efficient,
   title={Low-Altitude Dynamics of Very Flexible Aircraft},
   author={Del Carre, Alfonso and Palacios, Rafael},
-  booktitle={AIAA Scitech 2019 Forum},
+  booktitle={AIAA SciTech 2019 Forum},
   address={San Diego, California},
   year={2019},
   publisher = {American Institute of Aeronautics and Astronautics},
