### 2019-12-09 54c50a0
modify bib
