+++ docs/JOSS/paper.bib
@@ -345,14 +345,22 @@ year = {2012}
   title = {Dynamic Load Alleviation in Wake Vortex Encounters},
   journal = {Journal of Guidance,  Control,  and Dynamics}
  }
- @inproceedings{del2019efficient,
+
+@inproceedings{del2019efficient,
   title={Low-Altitude Dynamics of Very Flexible Aircraft},
   author={Del Carre, Alfonso and Palacios, Rafael},
-  booktitle={AIAA SciTech 2019 Forum},
   address={San Diego, California},
   year={2019},
   publisher = {American Institute of Aeronautics and Astronautics},
-  doi={10.2514/6.2019-2038}
+  doi={10.2514/6.2019-2038},
+  crossref={scitech19}
+}
+
+@proceedings{scitech19,
+  title={AIAA SciTech 2019 Forum},
+  publisher = {American Institute of Aeronautics and Astronautics},
+  venue={San Diego, California},
+  year={2019}
 }
 
 @article{melin2013tornado,
