### 2019-12-09 fa22ebf
Fixed capitalisations in bib
