+++ docs/JOSS/paper.bib
@@ -65,7 +65,7 @@
   publisher = {American Institute of Aeronautics and Astronautics},
   author = {Mark Drela},
   title = {Integrated simulation model for preliminary aerodynamic,  structural,  and control-law design of aircraft},
-  booktitle = {40th Structures,  Structural Dynamics,  and Materials Conference and Exhibit}
+  booktitle = {{40th Structures,  Structural Dynamics,  and Materials Conference and Exhibit}}
 }
 
 @PhdThesis{Shearer2006,
@@ -357,7 +357,7 @@ year = {2012}
 }
 
 @proceedings{scitech19,
-  title={AIAA SciTech 2019 Forum},
+  title={{AIAA SciTech 2019 Forum}},
   publisher = {American Institute of Aeronautics and Astronautics},
   venue={San Diego, California},
   year={2019}
@@ -411,7 +411,7 @@ eprint = {
   title={Low-Speed Aerodynamics},
   author={Katz, J. and Plotkin, A.},
   isbn={9781107717428},
-  series={Cambridge Aerospace Series},
+  series={{Cambridge Aerospace Series}},
   year={2001},
   publisher={Cambridge University Press}
 }
@@ -466,7 +466,7 @@ eprint = {https://arc.aiaa.org/doi/pdf/10.2514/6.2007-6703}
  @inproceedings{del2019ifasd,
   title={Nonlinear response of a very flexible aircraft under lateal gust},
   author={Del Carre, Alfonso and Teixeira, Patricia, and Cesnik, Carlos E. S. and Palacios, Rafael},
-  booktitle={IFASD 2019 Forum},
+  booktitle={{IFASD 2019 Forum}},
   address={Savannah, Georgia},
   year={2019},
   publisher = {},
