### 2019-12-09 a44f1f2
Documentation Improvements. Fixes #53
- SHARPy intro guide reworked into different sections.
- Moved media to `docs/source/_static`
- Disclaimer about incomplete documentation added.
