+++ docs/source/content/capabilities.md
@@ -1,0 +1,106 @@
+# Capabilities
+This is just the tip of the iceberg, possibilities are nearly endless and once you understand how SHARPys modularity
+works, you will be capable of running very complex simulations.
+
+## Very flexible aircraft nonlinear aeroelasticity
+
+The modular design of SHARPy allows to simulate complex aeroelastic cases involving very flexible aircraft. The
+structural solver supports very complex beam arrangements, while retaining geometrical nonlinearity. The UVLM solver
+features different wake modelling fidelities while supporting large lifting surface deformations in a native way.
+
+Among the problems studied, a few interesting ones, in no particular order are:
+
+* Catapult take off of a very flexible aircraft analysis [\[Paper\]](https://arc.aiaa.org/doi/abs/10.2514/6.2019-2038). 
+In this type of simulations, a PID controller was used in order to enforce displacements and velocities in a number of
+structural nodes (the clamping points). Then, several take off strategies were studied in order to analyse the influence
+of the structural stiffness in this kind of procedures. This case is a very good example of the type of problems where
+nonlinear aeroelasticity is essential.
+ 
+![CATO](../_static/capabilities/hale_cruise.png) 
+
+* Flight in full 3D atmospheric boundary layer (to be published). A very flexible aircraft is flown immersed in a
+turbulent boundary layer obtained from HPC LES simulations. The results are compared against simpler turbulence models
+such as von Karman and Kaimal. Intermittency and coherence features in the LES field are absent or less remarkable in
+the synthetic turbulence fields.
+
+![Turb](../_static/capabilities/hale_turb.jpeg)
+
+* Lateral gust reponse of a realistic very flexible aircraft. For this problem (to be published), a realistic very
+flexible aircraft (University of Michigan X-HALE) model has been created in SHARPy and validated against their own
+aeroelastic solver for static and dynamic cases. A set of vertical and lateral gust responses have been simulated.
+
+![XHALE](../_static/capabilities/xhale.png)
+
+## Wind turbine aeroelasticity
+
+SHARPy is suitable to simulate wind turbine aeroelasticity. 
+
+On the structural side, it accounts for material anisotropy which is needed to characterize composite blades and for
+geometrically non-linear deformations observed in current blades due to the increasing length and flexibility. Both
+rigid and flexible simulations can be performed and the structural modes can be computed accounting for rotational
+effects (Campbell diagrams). The rotor-tower interaction is modelled through a multibody approach based on the theory
+of Lagrange multipliers. Finally, he tower base can be fixed or subjected to prescribed linear and angular velocities.
+
+On the aerodynamic side, the use of potential flow theory allows the characterization of flow unsteadiness at a
+reasonable computational cost. Specifically, steady and dynamic simulations can be performed. The steady simulations
+are carried out in a non-inertial frame of reference linked to the rotor under uniform steady wind with the assumption
+of prescribed helicoidal wake. On the other hand, dynamic simulations can be enriched with a wide variety of incoming
+winds such as shear and yaw. Moreover, the wake shape can be freely computed under no assumptions accounting for
+self-induction and wake expansion or can be prescribed to an helicoidal shape for computational efficiency.
+
+![Wind Turbine](../_static/capabilities/turbine.png)
+
+## Model Order Reduction
+Numerical models of physical phenomena require fine discretisations to show convergence and agreement with their real
+counterparts, and, in the case of SHARPy's aeroelastic systems, hundreds of thousands of states are not an uncommon
+encounter. However, modern hardware or the use of these models for other applications such as controller synthesis may limit
+their size, and we must turn to model order reduction techniques to achieve lower dimensional representations that can
+then be used.
+
+SHARPy offers several model order reduction methods to reduce the initially large system to a lower dimension,
[…]
