### 2019-12-10 c703a90
Typos fixed
