### 2019-12-10 c5eb46b
Added settings options for single sided Krylov ROM
