### 2019-12-10 7447ed9
Minor changes in linear docs
