### 2019-12-10 b2e70ee
Update sharpy/rom/utils/librom.py
