### 2019-12-10 da9394f
Following from #57: Updated capabilities in readme
