### 2019-12-10 86f45c7
Removes vortex core suggestion from debug guide
Co-Authored-By: Alfonso del Carre <fonsocarre@users.noreply.github.com>
