### 2019-12-10 633ec59
Apply suggestions from code review
Co-Authored-By: Alfonso del Carre <fonsocarre@users.noreply.github.com>
