+++ sharpy/postproc/asymptoticstability.py
@@ -219,10 +219,10 @@ class AsymptoticStability(BaseSolver):
         ulb, uub, num_u = self.settings['velocity_analysis']
 
         if self.settings['print_info']:
-            cout.cout_wrap('Velocity Asymptotic Stability Analysis')
+            cout.cout_wrap('Velocity Asymptotic Stability Analysis', 1)
             cout.cout_wrap('Initial velocity: %.2f m/s' % ulb, 1)
             cout.cout_wrap('Final velocity: %.2f m/s' % uub, 1)
-            cout.cout_wrap('Number of evaluations: %g' %num_u, 1)
+            cout.cout_wrap('Number of evaluations: %g' % num_u, 1)
 
         u_inf_vec = np.linspace(ulb, uub, int(num_u))
 
