### 2019-12-10 00f104d
New test case for librom. Removed test from actual module
