### 2019-12-10 09290ac
Further code review suggestions added
