+++ sharpy/postproc/frequencyresponse.py
@@ -55,15 +55,6 @@ class FrequencyResponse(solver_interface.BaseSolver):
     settings_table = settings_utils.SettingsTable()
     __doc__ += settings_table.generate(settings_types, settings_default, settings_description)
 
-    # settings_types['plot_type'] = 'str'
-    # settings_default['plot_type'] = 'bode'
-    #
-    # settings_types['inputs'] = 'list(int)'
-    # settings_default['inputs'] = []
-    #
-    # settings_types['outputs'] = 'list(int)'
-    # settings_default['outputs'] = []
-
     def __init__(self):
 
         self.settings = None
@@ -138,21 +129,25 @@ class FrequencyResponse(solver_interface.BaseSolver):
                 compute_fom = True
 
         if (self.settings['compute_fom'].value and self.settings['load_fom'] == '') or compute_fom:
-            cout.cout_wrap('Computing frequency response...')
-            cout.cout_wrap('Full order system:', 1)
+            if self.settings['print_info']:
+                cout.cout_wrap('Computing frequency response...')
+                cout.cout_wrap('Full order system:', 1)
             t0fom = time.time()
             Y_freq_fom = self.ss.freqresp(self.wv)
             tfom = time.time() - t0fom
             self.save_freq_resp(self.wv, Y_freq_fom, 'fom')
-            cout.cout_wrap('\tComputed the frequency response of the full order system in %f s' % tfom, 2)
+            if self.settings['print_info']:
+                cout.cout_wrap('\tComputed the frequency response of the full order system in %f s' % tfom, 2)
 
         if self.ssrom is not None:
-            cout.cout_wrap('Computing frequency response...')
-            cout.cout_wrap('Reduced order system:', 1)
+            if self.settings['print_info']:
+                cout.cout_wrap('Computing frequency response...')
+                cout.cout_wrap('Reduced order system:', 1)
             t0rom = time.time()
             Y_freq_rom = self.ssrom.freqresp(self.wv)
             trom = time.time() - t0rom
-            cout.cout_wrap('\tComputed the frequency response of the reduced order system in %f s' % trom, 2)
+            if self.settings['print_info']:
+                cout.cout_wrap('\tComputed the frequency response of the reduced order system in %f s' % trom, 2)
             self.save_freq_resp(self.wv, Y_freq_rom, 'rom')
 
             if Y_freq_fom is not None:
@@ -209,8 +204,9 @@ class FrequencyResponse(solver_interface.BaseSolver):
             warnings.warn('Matplotlib not found - skipping plot')
 
     def load_frequency_data(self):
-        cout.cout_wrap('Loading frequency response from:')
-        cout.cout_wrap('\t%s' % self.settings['load_fom'], 1)
+        if self.settings['print_info']:
+            cout.cout_wrap('Loading frequency response from:')
+            cout.cout_wrap('\t%s' % self.settings['load_fom'], 1)
         Y_freq_fom = np.zeros((self.ss.outputs, self.ss.inputs, len(self.wv)), dtype=complex)
         for m in range(self.ss.inputs):
             for p in range(self.ss.outputs):
@@ -261,416 +257,3 @@ def error_between_signals(sig1, sig2, wv, sig_title=''):
 
     return max_error
 
-    # def plot_frequency_response(self, kv, Y_freq_ss, Y_freq_rom, interp_frequencies):
-    #
-    #     nstates = self.ss.states
-    #     if self.rom is not None and Y_freq_rom is not None:
-    #         rstates = self.rom.ssrom.states
[…]
