### 2019-12-10 ae27363
Improvements from code review
