+++ sharpy/rom/balanced.py
@@ -36,11 +36,14 @@ def bal_rom(arg):
 
 class BaseBalancedRom(metaclass=ABCMeta):
 
+    print_info = False
+
     def initialise(self, in_settings=None):
         pass
 
     def run(self, ss):
         pass
+
 
 @bal_rom
 class Direct(BaseBalancedRom):
@@ -55,6 +58,10 @@ class Direct(BaseBalancedRom):
     settings_types['tune'] = 'bool'
     settings_default['tune'] = True
     settings_description['tune'] = 'Tune ROM to specified tolerance'
+
+    settings_types['use_schur'] = 'bool'
+    settings_default['use_schur'] = False
+    settings_description['use_schur'] = 'Use Schur decomposition during build'
 
     settings_types['rom_tolerance'] = 'float'
     settings_default['rom_tolerance'] = 1e-2
@@ -88,7 +95,8 @@ class Direct(BaseBalancedRom):
         settings.to_custom_types(self.settings, self.settings_types, self.settings_default, self.settings_options)
 
     def run(self, ss):
-        cout.cout_wrap('Reducing system using a Direct balancing method...')
+        if self.print_info:
+            cout.cout_wrap('Reducing system using a Direct balancing method...')
         t0 = time.time()
         A, B, C, D = ss.get_mats()
 
@@ -100,7 +108,7 @@ class Direct(BaseBalancedRom):
         except AttributeError:
             dtsystem = False
 
-        S, T, Tinv = librom.balreal_direct_py(A, B, C, DLTI=dtsystem)
+        S, T, Tinv = librom.balreal_direct_py(A, B, C, DLTI=dtsystem, Schur=self.settings['use_schur'])
 
         Ar = T.dot(A.dot(Tinv))
         Br = T.dot(B)
@@ -112,7 +120,8 @@ class Direct(BaseBalancedRom):
             ss_bal = libss.ss(Ar, Br, Cr, D)
 
         t1 = time.time()
-        cout.cout_wrap('\t...completed balancing in %.2fs' % (t1-t0), 1)
+        if self.print_info:
+            cout.cout_wrap('\t...completed balancing in %.2fs' % (t1-t0), 1)
 
         if self.settings['tune']:
             cout.cout_wrap('\t\tTuning ROM to specified tolerance...', 2)
@@ -125,7 +134,8 @@ class Direct(BaseBalancedRom):
                                     convergence=self.settings['convergence'],
                                     method=self.settings['reduction_method'])
             if librom.check_stability(ssrom.A, dt=True):
-                cout.cout_wrap('ROM by direct balancing is stable')
+                if self.settings['print_info']:
+                    cout.cout_wrap('ROM by direct balancing is stable')
             t2 = time.time()
             cout.cout_wrap('\t...completed reduction in %.2fs' % (t2-t0), 1)
             return ssrom
@@ -295,7 +305,7 @@ class Balanced(rom_interface.BaseRom):
     Supported algorithms:
         * Direct balancing :class:`.Direct`
 
-        * Iterative balancing
+        * Iterative balancing :class:`.Iterative`
 
         * Frequency limited balancing :class:`.FrequencyLimited`
 
@@ -305,17 +315,23 @@ class Balanced(rom_interface.BaseRom):
     settings_types = dict()
     settings_default = dict()
     settings_description = dict()
+    settings_options = dict()
+
+    settings_types['print_info'] = 'bool'
+    settings_default['print_info'] = True
+    settings_description['print_info'] = 'Write output to screen'
 
     settings_types['algorithm'] = 'str'
     settings_default['algorithm'] = ''
     settings_description['algorithm'] = 'Balanced realisation method'
[…]
