### 2019-12-10 21c7e28
New test for the execution of the Balancing roms
