+++ tests/linear/rom/test_rom_framework.py
@@ -1,0 +1,262 @@
+import numpy as np
+import os
+import unittest
+import cases.templates.flying_wings as wings
+import sharpy.sharpy_main
+
+
+class TestROMFramework(unittest.TestCase):
+    """
+    Test verifyies the execution of the balancing ROMs (i.e. checks that everything runs rather than checking
+    against a benchmark case
+    """
+    
+    def setUp(self):
+        # Problem Set up
+        u_inf = 1.
+        alpha_deg = 0.
+        rho = 1.02
+        num_modes = 4
+
+        # Lattice Discretisation
+        M = 4
+        N = 8
+        M_star_fact = 1
+
+        # Linear UVLM settings
+        integration_order = 2
+        remove_predictor = False
+        use_sparse = False
+
+        # Case Admin - Create results folders
+        case_name = 'goland_cs'
+        case_nlin_info = 'M%dN%dMs%d_nmodes%d' % (M, N, M_star_fact, num_modes)
+        case_name += case_nlin_info
+
+        self.route_test_dir = os.path.abspath(os.path.dirname(os.path.realpath(__file__)))
+        fig_folder = self.route_test_dir + '/figures/'
+        os.makedirs(fig_folder, exist_ok=True)
+
+        # SHARPy nonlinear reference solution
+        ws = wings.Goland(M=M,
+                          N=N,
+                          Mstar_fact=M_star_fact,
+                          u_inf=u_inf,
+                          alpha=alpha_deg,
+                          rho=rho,
+                          sweep=0,
+                          physical_time=2,
+                          n_surfaces=2,
+                          route=self.route_test_dir + '/cases',
+                          case_name=case_name)
+
+        ws.gust_intensity = 0.01
+        ws.sigma = 1
+        ws.horseshoe = True
+
+        ws.clean_test_files()
+        ws.update_derived_params()
+        ws.set_default_config_dict()
+
+        ws.generate_aero_file()
+        ws.generate_fem_file()
+
+        ws.config['SHARPy'] = {
+            'flow':
+                ['BeamLoader', 'AerogridLoader',
+                 'StaticCoupled',
+                 'AerogridPlot',
+                 'BeamPlot',
+                 'Modal',
+                 'LinearAssembler',
+                 'FrequencyResponse',
+                 ],
+            'case': ws.case_name, 'route': ws.route,
+            'write_screen': 'off', 'write_log': 'on',
+            'log_folder': self.route_test_dir + '/output/' + ws.case_name + '/',
+            'log_file': ws.case_name + '.log'}
+
+        ws.config['BeamLoader'] = {
+            'unsteady': 'off',
+            'orientation': ws.quat}
+
+        ws.config['AerogridLoader'] = {
+            'unsteady': 'off',
+            'aligned_grid': 'on',
+            'mstar': 1,
+            'freestream_dir': ws.u_inf_direction
+        }
+
+        ws.config['StaticUvlm'] = {
+            'rho': ws.rho,
+            'velocity_field_generator': 'SteadyVelocityField',
+            'velocity_field_input': {
+                'u_inf': ws.u_inf,
+                'u_inf_direction': ws.u_inf_direction},
+            'rollup_dt': ws.dt,
+            'print_info': 'on',
+            'horseshoe': 'off',
+            'num_cores': 4,
+            'n_rollup': 0,
+            'rollup_aic_refresh': 0,
+            'rollup_tolerance': 1e-4}
+
+        ws.config['StaticCoupled'] = {
+            'print_info': 'on',
+            'max_iter': 200,
+            'n_load_steps': 1,
+            'tolerance': 1e-10,
+            'relaxation_factor': 0.,
+            'aero_solver': 'StaticUvlm',
+            'aero_solver_settings': {
+                'rho': ws.rho,
+                'print_info': 'off',
[…]
