+++ docs/_sources/content/structure.rst.txt
@@ -1,2 +1,0 @@
-Structural Models
------------------
