+++ docs/_sources/content/generators.rst.txt
@@ -1,4 +1,0 @@
-Generators
-----------
-
-
