+++ docs/_sources/content/utils/algebra_content.rst.txt
@@ -1,12 +1,0 @@
-The ``Algebra`` package
------------------------
-This package includes routines used often during geometry generation.
-
-Tangent vector calculation
-++++++++++++++++++++++++++
-
-
-.. automodule:: sharpy.utils.algebra
-    :members:
-
-
