+++ docs/_sources/content/utilities.rst.txt
@@ -1,8 +1,0 @@
-Utilities
-=========
-
-The following ``utilities`` modules are provided in ``SHARPy``
-
-.. toctree::
-    utils/algebra_content
-    utils/h5utils
