+++ docs/_sources/content/solvers.rst.txt
@@ -1,42 +1,0 @@
-``SHARPy`` Solvers
-------------------
-
-``SHARPy`` has the following solvers available for use:
-
-IO of Models
-++++++++++++
-
-.. autoclass:: sharpy.solvers.beamloader.BeamLoader
-    :members:
-
-.. autoclass:: sharpy.solvers.aerogridloader.AerogridLoader
-    :members:
-
-
-
-Structural Solvers
-++++++++++++++++++
-
-Aerodynamic Solvers
-+++++++++++++++++++
-
-
-Aeroelastic Solvers
-+++++++++++++++++++
-.. autoclass:: sharpy.solvers.dynamiccoupled.DynamicCoupled
-    :members:
-
-.. autoclass:: sharpy.solvers.dynamicprescribedcoupled.DynamicPrescribedCoupled
-    :members:
-
-Modal Solvers
-+++++++++++++
-
-.. autoclass:: sharpy.solvers.modal.Modal
-    :members:
-
-Flight Solvers
-++++++++++++++
-.. autoclass:: sharpy.solvers.statictrim.StaticTrim
-    :members:
-
