+++ _posts/2019-12-10-mass-stiffness-effect-on-damping.md
@@ -1,0 +1,8 @@
+---
+layout: post
+title: 'Mass and stiffness effect on damping'
+date: 2019-12-10
+author: Norberto Goizueta
+---
+
+While testing a 
