+++ docs/_sources/content/utils/h5utils.rst.txt
@@ -1,5 +1,0 @@
-HDF5 Utilities Module
-=====================
-
-.. automodule:: sharpy.utils.h5utils
-    :members:
