+++ docs/_sources/content/aero.rst.txt
@@ -1,5 +1,0 @@
-Aerodynamic Models
-------------------
-
-.. autoclass:: sharpy.aero.models.aerogrid.Aerogrid
-    :members:
