+++ docs/_sources/index.rst.txt
@@ -1,35 +1,0 @@
-.. SHARPy documentation master file, created by
-   sphinx-quickstart on Wed Oct 19 16:40:48 2016.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-SHARPy documentation
-====================
-
-
-
-Contents:
----------
-
-.. toctree::
-   :maxdepth: 2
-
-   content/installation
-   content/presharpy
-   content/run_sharpy
-   content/aero
-   content/generators
-   content/structure
-   content/solvers
-   content/postproc
-   content/utilities
-   content/test_cases
-
-
-Indices and tables
-------------------
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
-
