### 2019-12-11 10e7b95
deg2rad as constant
