### 2019-12-11 63b4789
cleanup template_wt
