+++ cases/templates/template_wt.py
@@ -7,22 +7,22 @@ Notes:
     To load this library: import cases.templates.template_wt as template_wt
 """
 
-import sharpy.utils.generate_cases as gc
+
 import pandas as pd
 import numpy as np
 import scipy.interpolate as scint
 import math
-import os
+import sys
+
+import sharpy.utils.generate_cases as gc
 import sharpy.utils.algebra as algebra
 import sharpy.utils.h5utils as h5
 from sharpy.utils.constants import deg2rad
 
 
-
 ######################################################################
 # AUX FUNCTIONS
 ######################################################################
-
 def create_node_radial_pos_from_elem_centres(root_elem_centres_tip, num_node, num_elem, num_node_elem):
     """
     create_node_radial_pos_from_elem_centres
@@ -75,1004 +75,58 @@ def create_blade_coordinates(num_node, node_r, node_y, node_z):
     Returns:
         coordinates (np.array): nodes coordinates
     """
-    coordinates = np.zeros((num_node,3),)
-    coordinates[:,0] = node_r
-    coordinates[:,1] = node_y
-    coordinates[:,2] = node_z
+    coordinates = np.zeros((num_node, 3),)
+    coordinates[:, 0] = node_r
+    coordinates[:, 1] = node_y
+    coordinates[:, 2] = node_z
     return coordinates
 
-######################################################################
-# FROM EXCEL TYPE 01
-######################################################################
-def generate_from_excel_type01(chord_panels,
-                                  rotation_velocity,
-                                  pitch,
-                                  excel_file_name= 'database_type01.xlsx',
-                                  excel_sheet_structural_blade = 'structural_blade',
-                                  excel_sheet_aero_blade = 'aero_blade',
-                                  excel_sheet_airfoil_coord = 'airfoil_coord',
-                                  excel_sheet_rotor = 'rotor_parameters',
-                                  excel_sheet_structural_tower = 'structural_tower',
-                                  excel_sheet_nacelle = 'structural_nacelle',
-                                  m_distribution = 'uniform',
-                                  n_points_camber = 100,
-                                  tol_remove_points = 1e-3):
-
-    """
-    generate_wt_from_excel_type01
-
-    Function needed to generate a wind turbine from an excel database of type 01 (FAST format)
-
-    Args:
-    	  chord_panels (int): Number of panels on the blade surface in the chord direction
-          rotation_velocity (float): Rotation velocity of the rotor
-          pitch (float): pitch angle in degrees
-          excel_file_name (str):
-          excel_sheet_structural_blade (str):
-          excel_sheet_aero_blade (str):
-          excel_sheet_airfoil_coord (str):
-          excel_sheet_rotor (str):
-          excel_sheet_structural_tower (str):
-          excel_sheet_nacelle (str):
-          m_distribution (str):
-          n_points_camber (int): number of points to define the camber of the airfoil,
-          tol_remove_points (float): maximum distance to remove adjacent points
-
-    Returns:
-        wt (sharpy.utils.generate_cases.AeroelasticInfromation): Aeroelastic infrmation of the wind turbine
[…]
