### 2019-12-11 a5e5891
cleanup and docs
