+++ sharpy/utils/generate_cases.py
@@ -6,6 +6,10 @@ This library provides functions and classes to help in the definition of SHARPy 
 Examples:
 
     tests in: tests/utils/generate_cases
+    examples: test/coupled/multibody/fix_node_velocity_wrtG/test_fix_node_velocity_wrtG
+              test/coupled/multibody/fix_node_velocity_wrtA/test_fix_node_velocity_wrtA
+              test/coupled/multibody/double_pendulum/test_double_pendulum_geradin
+              test/coupled/prescribed/WindTurbine/test_rotor
 
 Notes:
 
@@ -14,17 +18,17 @@ Notes:
 """
 
 import numpy as np
-import sharpy.utils.algebra as algebra
-import sharpy.utils.solver_interface as solver_interface
-import sharpy.utils.generator_interface as generator_interface
-import sharpy.structure.utils.lagrangeconstraints as lagrangeconstraints
-# from sharpy.utils.solver_interface import solver, dict_of_solvers
 import h5py as h5
 import os
 import sys
 import pandas as pd
 import scipy.integrate
 from copy import deepcopy
+
+import sharpy.utils.algebra as algebra
+import sharpy.utils.solver_interface as solver_interface
+import sharpy.utils.generator_interface as generator_interface
+import sharpy.structure.utils.lagrangeconstraints as lagrangeconstraints
 
 
 ######################################################################
@@ -37,16 +41,16 @@ def get_airfoil_camber(x, y, n_points_camber):
     Define the camber of an airfoil based on its coordinates
 
     Args:
-    	x (np.array): x coordinates of the airfoil surface
-    	y (np.array): y coordinates of the airfoil surface
+        x (np.array): x coordinates of the airfoil surface
+        y (np.array): y coordinates of the airfoil surface
         n_points_camber (int): number of points to define the camber line
 
     Returns:
         camber_x (np.array): x coordinates of the camber line
-    	camber_y (np.array): y coordinates of the camber line
+        camber_y (np.array): y coordinates of the camber line
 
     Notes:
-    	The x and y vectors are expected in XFOIL format: TE - suction side - LE - pressure side - TE
+        The x and y vectors are expected in XFOIL format: TE - suction side - LE - pressure side - TE
 
     """
     # Returns the airfoil camber for a given set of coordinates (XFOIL format expected)
@@ -85,6 +89,7 @@ def get_airfoil_camber(x, y, n_points_camber):
     # The function should be called as: camber_x, camber_y = get_airfoil_camber(x,y)
     return camber_x, camber_y
 
+
 def from_node_list_to_elem_matrix(node_list, connectivities):
     """
     from_node_list_to_elem_matrix
@@ -101,15 +106,18 @@ def from_node_list_to_elem_matrix(node_list, connectivities):
         connectivities (np.array): Connectivities between the nodes to form elements
 
     Returns:
-    	elem_matrix (np.array): Properties of the elements
+        elem_matrix (np.array): Properties of the elements
+
     """
+
     num_elem = len(connectivities)
     # TODO: change the "3" for self.num_node_elem
-    elem_matrix = np.zeros((num_elem,3), dtype=node_list.dtype)
+    elem_matrix = np.zeros((num_elem, 3), dtype=node_list.dtype)
     for ielem in range(num_elem):
         elem_matrix[ielem, :] = node_list[connectivities[ielem, :]]
 
     return elem_matrix
+
 
 def from_node_array_to_elem_matrix(node_array, connectivities):
     """
@@ -117,17 +125,37 @@ def from_node_array_to_elem_matrix(node_array, connectivities):
 
     Same as the previous function but with an array as input
     """
+
     num_elem = len(connectivities)
     prop_size = node_array.shape[1:]
     # TODO: change the "3" for self.num_node_elem
[…]
