### 2019-12-11 34bb640
update cases, test and examples wind turbines
