+++ tests/coupled/multibody/double_pendulum/test_double_pendulum_geradin.py
@@ -112,6 +112,7 @@ geradin_FoR1 = np.array([[0.00756934, 0.0266485],
                         [4.91188, 0.74549],
                         [4.98432, 0.797635]])
 
+
 class TestDoublePendulum(unittest.TestCase):
     """
     Validation of a double pendulum with a mass at each tip position
