+++ tests/utils/test_generate_cases.py
@@ -1,12 +1,8 @@
 import unittest
 import numpy as np
-import sys
 import os
 import shutil
-# import math
-# import pandas as pd
-#
-# import sharpy.utils.algebra as algebra
+
 import sharpy.utils.generate_cases as gc
 import cases.templates.template_wt as template_wt
 from sharpy.utils.constants import deg2rad
@@ -61,7 +57,7 @@ class TestGenerateCases(unittest.TestCase):
                                           chord_panels,
                                           rotation_velocity,
                                           pitch_deg,
-                                          excel_file_name = route + 'type02_db_NREL_5MW.xlsx',
+                                          excel_file_name = '../../docs/source/content/example_notebooks/source/type02_db_NREL5MW_v01.xlsx',
                                           excel_sheet_parameters = 'parameters',
                                           excel_sheet_structural_blade = 'structural_blade',
                                           excel_sheet_discretization_blade = 'discretization_blade',
