+++ docs/source/includes/utils/generate_cases/get_factor_geometric_progression.rst
@@ -1,0 +1,4 @@
+get_factor_geometric_progression
+--------------------------------
+
+.. automodule:: sharpy.utils.generate_cases.get_factor_geometric_progression
