### 2019-12-12 9c8db4d
Updated changelog and version badge
