+++ CHANGELOG.md
@@ -1,6 +1,50 @@
 # Changelog
 
-[Full Changelog](https://github.com/imperialcollegelondon/sharpy/compare/1.0.0...HEAD)
+## [1.1](https://github.com/imperialcollegelondon/sharpy/tree/1.1) (2019-12-12)
+
+[Full Changelog](https://github.com/imperialcollegelondon/sharpy/compare/v1.0.1...1.1)
+
+**Implemented enhancements:**
+
+- Improvements to Model Reduction [\#56](https://github.com/ImperialCollegeLondon/sharpy/pull/56) ([ngoiz](https://github.com/ngoiz))
+- Submodules and cmake build tools instead of Makefiles [\#52](https://github.com/ImperialCollegeLondon/sharpy/pull/52) ([fonsocarre](https://github.com/fonsocarre))
+- New settings check against valid options [\#39](https://github.com/ImperialCollegeLondon/sharpy/pull/39) ([ngoiz](https://github.com/ngoiz))
+- Default settings get type specified for the setting rather than their own type [\#34](https://github.com/ImperialCollegeLondon/sharpy/pull/34) ([ngoiz](https://github.com/ngoiz))
+
+**Fixed bugs:**
+
+- Default value not correct in documentation when it is a numpy type. [\#32](https://github.com/ImperialCollegeLondon/sharpy/issues/32)
+- Documentation for Postprocessors being skipped by Sphinx in RTD [\#30](https://github.com/ImperialCollegeLondon/sharpy/issues/30)
+
+**Closed issues:**
+
+- Minor documentation issues [\#53](https://github.com/ImperialCollegeLondon/sharpy/issues/53)
+- WindTurbine case generation script does not produce sharpy file [\#50](https://github.com/ImperialCollegeLondon/sharpy/issues/50)
+- Considerations for building SHARPy [\#47](https://github.com/ImperialCollegeLondon/sharpy/issues/47)
+- Installation fails on macOS with Intel compiler [\#46](https://github.com/ImperialCollegeLondon/sharpy/issues/46)
+- run\_theo\_freq.py fails in Docker container [\#37](https://github.com/ImperialCollegeLondon/sharpy/issues/37)
+- Compare to other competing software in JOSS paper [\#36](https://github.com/ImperialCollegeLondon/sharpy/issues/36)
+
+**Merged pull requests:**
+
+- Example wind turbine [\#58](https://github.com/ImperialCollegeLondon/sharpy/pull/58) ([ArturoMS13](https://github.com/ArturoMS13))
+- Small typo in README.md and updates to it [\#57](https://github.com/ImperialCollegeLondon/sharpy/pull/57) ([fonsocarre](https://github.com/fonsocarre))
+- Restructuring of A Short Guide to SHARPy [\#55](https://github.com/ImperialCollegeLondon/sharpy/pull/55) ([ngoiz](https://github.com/ngoiz))
+- JOSS Paper Minor typos fixed [\#54](https://github.com/ImperialCollegeLondon/sharpy/pull/54) ([ngoiz](https://github.com/ngoiz))
+- Update .solver.txt extension to .sharpy [\#51](https://github.com/ImperialCollegeLondon/sharpy/pull/51) ([ArturoMS13](https://github.com/ArturoMS13))
+- Fix typo in unittests using tearDowns instead of tearDown [\#49](https://github.com/ImperialCollegeLondon/sharpy/pull/49) ([ngoiz](https://github.com/ngoiz))
+- Bug fixes in installation docs [\#45](https://github.com/ImperialCollegeLondon/sharpy/pull/45) ([rafmudaf](https://github.com/rafmudaf))
[…]
