### 2019-12-12 5bd5eb8
Fixed X-HALE image address in JOSS paper
