### 2019-12-12 #1885 @ngoiz
Hi all,

We have now released SHARPy v1.1 which includes all the comments from the review process. It is archived in Zenodo https://zenodo.org/record/3572512 with DOI: 10.5281/zenodo.3572512

That is the version specific DOI. The concept DOI is 10.5281/zenodo.3531965

Thanks again for your reviews and suggestions!
