### 2019-12-12 #1885 @fonsocarre
I just wanted to thank you all for the constructive suggestions and the positive tone of the review process. SHARPy now is much easier to use than when we submitted. If there's some aero-related software and you need a reviewer, give me a shout, I'll be happy to contribute.
