### 2019-12-13 #1885 @labarba
We generally ask for the _title_ of the Zenodo archive to match the title of the JOSS paper, but you didn't make that change. Do you have any strong reasons?
